### 2025-01-04 #231 @pudeIko
PIVA Submission
Submitting Author: Wojciech Radoslaw Pudelko (@pudeIko)
All current maintainers: Wojciech Radoslaw Pudelko (@pudeIko)
Package Name: PIVA
One-Line Description of Package: Visualization and analysis toolkit for experimental data from Angle-Resolved Photoemission Spectroscopy (ARPES)
Repository Link:  https://github.com/pudeIko/piva
Version submitted: v2.3.1
EiC: TBD
Version accepted: TBD

---

PIVA (Photoemission Interface for Visualization and Analysis) is a GUI application designed for the interactive and intuitive exploration of large, image-like datasets. While it accommodates the visualization of any multidimensional data, its features are specifically optimized for researchers conducting Angle-Resolved Photoemission Spectroscopy (ARPES) experiments. In addition to numerous image processing tools and the ability to apply technique-specific corrections, PIVA includes an expanding library of functions and methods for detailed fitting and advanced spectral analysis.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

**Data extraction:** Within the ARPES community, it is common for each beamline and lab to use their own file formats and conventions, which means one often need a custom script to get everything into a common format. To handle these discrepancies, PIVA comes with a `data_loaders` module that converts them into a standardized `Dataset` object. The current version includes specific Dataloader classes implemented for numerous sources and beamlines around the world.

**Data visualization:** The package enables efficient and intuitive exploration of large, image-like datasets. It includes specialized interactive viewers designed to handle 2D, 3D, and 4D datasets, depending on the experimental mode or conditions under which they were collected.

Experimental physicists conducting ARPES measurements. The package provides a comprehensive framework addressing most of the experimenter's needs, including data extraction, inspection, validation, and detailed analysis.

Regarding software tailored for ARPES, two notable packages are [ARPES Python Tools](https://github.com/pranabdas/arpespythontools) and [PyARPES](https://github.com/chstan/arpes). However, they differ significantly from PIVA.

The visualization module in the former is limited to generating static plots and lacks any interactive features.

The latter is focused on post-processing and detailed analysis of the spectra, and is different in the following respects:
- interactive exploration and browsing through data is either restricted to 2D data, or conducted inside the Jupyter environment, which highly affects efficiency and makes working with multiple datasets simultaneously difficult.
- Viewers designed for 4D datasets are not implemented.
- PIVA's `data_loader` module contains richer library of data loading scripts for different light sources around the world.

Furthermore, PyARPES has not been maintained for several years.

#223 (@SimonMolinsky)

- [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.14599024](https://doi.org/10.5281/zenodo.14599024)
