### 2025-01-21 #231 @coatless
---
---

This is a very well done package, the documentation is nearly perfect with the detailed description of each GUI element and the ability to specify a custom data set. The only issue is the lack of _short_ walk through from data ingestion to modeling using a sample data set.

In addition, it would be highly beneficial to show a fully working custom data loader for a given ARPES file that has a non-standard format. The present example is too sparse.

https://piva.readthedocs.io/en/latest/dataloaders.html#writing-a-custom-dataloader

Please also fix the parameter labels under the API documentation for `DataBrowser.add_viewer_to_linked_list()`

https://piva.readthedocs.io/en/latest/modules/data_browser.html#data_browser.DataBrowser.add_viewer_to_linked_list
