### 2025-01-21 adaeaa5
minor adjustments to fitting functions
