### 2025-01-22 bd07252
adds a non-standard ARPES file to test configuration of custom DLs
