### 2025-01-23 4a5bfd9
add the showcase video and analysis examples to documentation
