### 2025-01-23 5e1a21d
update docs requirements
