+++ docs/requirements.txt
@@ -1,3 +1,5 @@
 sphinx
 sphinx_rtd_theme
 piva
+nbsphinx
+myst-nb
