### 2025-01-23 #231 @pudeIko
Hi @coatless,

Thanks for the great feedback and initial remarks! I've implemented the requested changes:

> The only issue is the lack of short walk through from data ingestion to modeling using a sample data set.

I've added a new section to the documentation with an example covering data extraction and basic analysis. I hope this aligns with what you had in mind.

> In addition, it would be highly beneficial to show a fully working custom data loader for a given ARPES file that has a non-standard format. The present example is too sparse.
> https://piva.readthedocs.io/en/latest/dataloaders.html#writing-a-custom-dataloader

I've prepared a simulated data file and updated the `CustomDataloader` class so users can test proper configuration and have a more robust template for their code.

> Please also fix the parameter labels under the API documentation for DataBrowser.add_viewer_to_linked_list()
> https://piva.readthedocs.io/en/latest/modules/data_browser.html#data_browser.DataBrowser.add_viewer_to_linked_list

Done!

I have a technical question. I've committed all changes to the `main` branch and bumped the version number. Is this the correct approach, or would it be better to keep the changes in a separate branch during the review?
