### 2025-01-25 b4c70cd
adds logo to README and makes small improvements in docs
