+++ docs/source/conf.py
@@ -58,7 +58,7 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
-html_logo = '../img/logo.png'
+html_logo = '../img/logo-docs.png'
 html_theme_options = {
     "logo_only": True,
 }
