### 2025-03-30 #231 @coatless
@pudeIko Thank you for your patience! We've secured an editor to further move the review along.

We're happy to announce that @crhea93 will be the editor for your submission.

For next step, we'll be working toward getting reviewers assigned. This step is detailed here:

https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#the-review-begins

I'll let @crhea93 introduce himself 🙂
