### 2025-03-31 #231 @crhea93
@pudeIko Good morning!
Just to introduce my self briefly -- I am a Ph.D. scientist at the Dragonfly Telescope working on photometric and spectroscopic  analysis. I'm very excited to help you through the process of a successful submission!

As @coatless mentioned, the first step is to procure two reviewers -- I  will be working on this diligently this week.

Please let me know if you have any questions or need any assistance throughout this process.
