### 2025-04-02 #231 @crhea93
## Editor response to review:

---

## Editor comments
:wave: Hi @jsdodge and @eigenbrot! Thank you for volunteering to review
for pyOpenSci! I look forward to a successful review!

The first thing you will do is fill out this pre-review survey. Once completed, please move on to the review itself following the review template outlined below.

Please let me know when you have completed your survey so that I can check it off here. You can either let me know in a tagged comment below or in a separate message on slack or via email.

Please get in touch with any questions or concerns! Your review is due: April 22nd, 2025
---

Reviewers: jsdodge, eigenbrot
Due date: 04/22/2025
