### 2025-04-19 #231 @jsdodge
Estimated hours spent reviewing: 8–10

---

##### Summary

This is a very nice package, @pudelko, congratulations! I don't do ARPES myself but I can see the value that this will have to the community.

I'm very sorry this took me so long. I went through the documentation and took notes on things that I found confusing or didn't work for me, which I am including below. I think most of these are fairly minor.

Overall, the documentation is clear and I was able to follow most sections without difficulty, although I was unable to use the Plot Tool or open JupyterLab from within the application. I think it would be helpful to have one or more examples that show how a typical analysis might proceed, so that a new user can see how the different GUI elements interact.

##### README
- This should include a short description that compares Piva to other packages in use by the ARPES community. It can be similar to what you already have in the JOSS paper.

##### Installation

- I get the following warning when running `piva`:

	piva/main.py:1: UserWarning: pkg_resources is deprecated as an
API. See https://setuptools.pypa.io/en/latest/pkg_resources.html.
The pkg_resources package is slated for removal as early as
2025-11-30. Refrain from using this package or pin to
Setuptools<81.

##### Tests
- Running `pytest` returned a message that it passed 4 tests and generated 21 warnings, which is different from the message given in the [documentation](https://piva.readthedocs.io/en/latest/installation.html#testing).

##### Data browser
- [Documentation](https://piva.readthedocs.io/en/latest/db.html) uses `CTRL-O`, macOS uses `CMD-O` or `⌘-O` (I don't think this is a big deal, so feel free to ignore it if you think fixing it makes the documentation harder to read).

##### Data viewers
- The viewer axes are not labeled, nor are the values for the sliders in the various tabs. Would it be a problem to do this? It might help new users.
- In the image of the 2D DataViewer shown in the [documentation](https://piva.readthedocs.io/en/latest/viewers.html#sec-utilities-panel), panel (d) has a different item order than in the actual application. The items are different in the 3D and 4D DataViewers, but their documentation tables point to the list in the 2D DataViewer table.
- In the 2D DataViewer, selecting "Do it", then "Reset" changes the button name to "Do curvature".
- The 3D DataViewer documentation images show "Orientate" while the application shows "Orient".
- Selecting the "open in PIT" button in the 2D DataViewer does not issue a warning, but in the 3D DataViewer panel it does. It opens a box with the warning, "Note: PIT is a third-party package and may not work properly with Python versions above 3.8."
- The table in the 3D DataViewer [documentation](https://piva.readthedocs.io/en/latest/viewers.html#sec-3d-viewer) does not match the panels shown in the figure.
- Selecting the "start JL session" from the "File" tab does not start JupyterLab in my installation. It does open a system file browser window, but it's not clear to me what I should select to open from this window. Files with `.ipynb` extensions are visible but I can not select them for opening.
- The "File" tab for the 2D DataViewer has an annotation "create experimental logbook" but the table lists "open experimental logbook" in the "Functionality" column (the description says it creates it, though).

##### Fitters
- MDC Fitter returns error when saving results:

  	File "/Users/jsd/Library/CloudStorage/OneDrive-	SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/	fitters.py", line 1057, in save_fit_results
   for result in self.fit_results:
	TypeError: 'NoneType' object is not iterable

  	File "/Users/jsd/Library/CloudStorage/OneDrive-	SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/	fitters.py", line 1063, in save_fit_results
    for entry in self.fit_results:
	TypeError: 'NoneType' object is not iterable

	- It's not obvious where the fit results are stored

##### Plot tool
- It is unclear to me how to use the Plot Tool.
