### 2025-04-21 #231 @eigenbrot
- **Currently says "TBD"**

      - **See below for a more detailed description**
      - **It appears that the CI linting step is failing, but being ignored**

---

##### Summary

The PIVA package is an incredibly well realized toolkit for analyzing multidimensional data, specifically data from ARPES experiments. I don't personally work with ARPES data, but after reading the thorough documentation and playing with PIVA for a few hours I felt I had gained a good understanding of the common data viewing and analysis techniques of that field.

PIVA is undoubtedly a very useful tool for analyzing complex data, and a great addition to the pyOpenSci community. PIVA's representation of, and interaction with, 3D data is so natural and intuitive that I wrote my own toy `DataLoader`subclass to view 3D FITS data (a common format in astronomy) and it works great! I *love* the way binning is represented in the Data Viewer.

The biggest weakness from a user-facing experience is a gap in the documentation regarding the final steps of implementing custom data loaders and GUI plugins. This could be easily remedied with more detailed documentation and examples.

The biggest weakness from a packaging perspective is how the tests are implemented and the test coverage in general.

##### Specific Comments

###### Custom Loaders/Plugins

1. The construction of either a `Dataloader` subclass or a custom QT widget is generally well described, but how they are imported into PIVA through either `DataloaderImporter` or `PluginImporter` is less clear. Examples of these `*Importer` classes are provided in external files, but I would really like to see a detailed description of the parameters and methods of these classes. They are *not* mentioned in the otherwise-complete API documentation. A suggestion would be to write [abstract](https://docs.python.org/3/library/abc.html) bases of these `*Importer` classes that clearly define the interface.

2. The fact that a `Dataloader` subclass *must* contain a `load_data()` method is a good indication that `Dataloader` should probably be an [abstract base class](https://docs.python.org/3/library/abc.html).

###### Test System

1. Per the documentation, PIVA tests are run by calling the test modules directly, which starts a `pytest` session. This felt weird to me; I think most people are used to simply running `pytest` in the project root, which *does* work for PIVA. Running `pytest` is the method for testing described in CONTRIBUTING.md.

2. `pytest` is not listed in the package requirements table. There is also no "test" pip extra.

3. Test coverage for the GUI functionality seems high, but the rest of the package could use more tests. In particular, there don't seem to be any tests of the `working_procedures` module.

4. Despite the presence of a "tox.ini" file, simply running `tox` locally fails. (I suspect `tox` is only used for CI builds, so maybe this doesn't matter)

###### Use of Older Frameworks

There are few places in the code where older methods and frameworks could be replaced with more modern alternatives that would improve the maintainability and robustness of the package. In particular:

1. The `Dataset` class inherits from python's `Namespace` class, which allows for properties to be accessing with a nice `ds.property` syntax. I would highly recommend moving to using either [python dataclasses](https://docs.python.org/3/library/dataclasses.html) or [pydantic](https://docs.pydantic.dev/latest/) `BaseModels`, both of which are specifically designed for this type of interaction. Pydantic's `BaseModel` also offers a lot of useful type-validation functionality that may be particularly useful because the internals of the `Dataset` are tightly constrained to how that class is used by other code.

2. In a few places (e.g., Notebooks created by the "touch" button in the Data Viewer) file paths are concatenated with a `+` operator. At the very least I would recommend using `os.path.join`, or (even better) `pathlib.Path` objects.

###### Documentation

Overall the documentation is extremely good. My comments are minor:

1. PIVA's [pypi page](https://pypi.org/project/piva/) does not contain a "Documentation" link to the readthedocs site. This is a shame because the documentation is so good!

2. I think the ["What is `piva`"](https://piva.readthedocs.io/en/latest/intro.html#what-is-piva) section would be more impactful if the multiple strengths/features of `piva` were given their own subheadings.

3. The ["Implemented Dataloaders"](https://piva.readthedocs.io/en/latest/dataloaders.html#implemented-dataloaders) table is great, and would be even more great if the individual data loaders were links to the respective class in the API documentation.

4. On the ["Data Ingestion and Analysis"](https://piva.readthedocs.io/en/latest/notebooks/analysis.html) page it would be very useful to have an example dataset (the same one used in the example code) so a user can follow along directly. I was eventually able to run the notebook with one of the "test/data" datasets; perhaps elevating one or more of these datasets to an "examples" folder would make their utility more obvious.

###### Misc

* Data Viewer line plots show nothing if `NaN` values are present. The images still render correctly.
* `astropy` is listed as a dependency, but I can't find its usage anywhere
* Related, the use of [`astropy.units`](https://docs.astropy.org/en/stable/units/index.html) might be helpful. It's also possible the entire [constants](https://piva.readthedocs.io/en/latest/modules/constants.html) module could be replaced with [`astropy.constants`](https://docs.astropy.org/en/stable/constants/index.html).
* When Saving data to a pickle (via the "save" button in the Data Viewer) it would be nice to be able to specify the save directory.
* I found `db` as the script name to be a little confusing. I kept expecting it to be `piva`.
* I was expecting the "touch" and "start JL session" buttons to populate a notebook with correct paths for the dataset loaded into the Data Viewer. Instead it looks like they just copied a generic template to the given directory.
* It looks like `ruff` errors are being ignored in CI
* In tiled window manager, content in windows gets truncated if the window is small enough
* Any opened Jupyter sessions (created with the "start JL session" button) persist after closing all Piva windows.
* Installation was painless with multiple methods, even those not described in the documentation. FWIW, PIVA appears to work correctly (and the tests pass) for all python versions between 3.10 and 3.13.
* When clicking the "open in PIT" button the resulting PIT window displays a terminal with the following error in python 3.11 (but not 3.10):
```TypeError                                 Traceback (most recent call last)
----> 1 get_ipython().run_line_magic('pylab', 'qt')

File ~/micromamba/envs/piva11/lib/python3.11/site-packages/IPython/core/interactiveshell.py:2481, in InteractiveShell.run_line_magic(self, magic_name, line, _stack_depth)
   2479     kwargs['local_ns'] = self.get_local_scope(stack_depth)
   2480 with self.builtin_trap:
-> 2481     result = fn(*args, **kwargs)
   2483 # The code below prevents the output from being displayed
   2484 # when using magics with decorator @output_can_be_silenced
   2485 # when the last Python token in the expression is a ';'.
   2486 if getattr(fn, magic.MAGIC_OUTPUT_CAN_BE_SILENCED, False):

File ~/micromamba/envs/piva11/lib/python3.11/site-packages/IPython/core/magics/pylab.py:159, in PylabMagics.pylab(self, line)
    156     # invert no-import flag
    157     import_all = not args.no_import_all
--> 159 gui, backend, clobbered = self.shell.enable_pylab(args.gui, import_all=import_all)
    160 self._show_matplotlib_backend(args.gui, backend)
    161 print(
    162     "%pylab is deprecated, use %matplotlib inline and import the required libraries."
    163 )

File ~/micromamba/envs/piva11/lib/python3.11/site-packages/ipykernel/inprocess/ipkernel.py:200, in InProcessInteractiveShell.enable_pylab(self, gui, import_all, welcome_message)
    198 if not gui:
    199     gui = self.kernel.gui
--> 200 return super().enable_pylab(gui, import_all, welcome_message)

TypeError: InteractiveShell.enable_pylab() takes from 1 to 3 positional arguments but 4 were given
```
