### 2025-05-13 5a25a5e
adds test extras to toml file
