+++ pyproject.toml
@@ -18,7 +18,6 @@ classifiers = [
     "Programming Language :: Python :: 3",
 ]
 dependencies = [
-    "astropy>=4.2",
     "data_slicer>=1.0.2",
     "h5py>=3.11.0",
     "igor2>=0.5",
@@ -37,6 +36,9 @@ dependencies = [
     "tqdm>=4.56.0",
     "typing>=3.7",
 ]
+
+[project.optional-dependencies]
+test = ["pytest", "tox"]
 
 [project.scripts]
 db = "piva.main:db"
