+++ piva/data_viewer_3d.py
@@ -1941,7 +1941,8 @@ class DataViewer3D(QtWidgets.QMainWindow):
         else:
             pass
 
-        dl.dump(dataset, (savedir + fname), force=True)
+        # dl.dump(dataset, (savedir + fname), force=True)
+        dl.dump(dataset, os.path.join(savedir, fname), force=True)
 
     def load_corrections(self, data_set: Dataset) -> None:
         """
