### 2025-05-14 bba49a7
adds directory selection when saving files to pickle
