+++ piva/data_viewer_4d.py
@@ -1099,7 +1099,9 @@ class DataViewer4D(QtWidgets.QMainWindow):
 
         # energy axis
         dataset = self.scan
-        savedir = self.fname[:-len(self.title)]
+        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
+            self, 'Select Directory', self.fname[:-len(self.title)]))
+        # savedir = self.fname[:-len(self.title)]
         file_selection = True
         # Prepare a filename with the .p suffix
         init_fname = '.'.join(self.title.split('.')[:-1] + ['p'])
@@ -1129,7 +1131,6 @@ class DataViewer4D(QtWidgets.QMainWindow):
             else:
                 file_selection = False
 
-        # dl.dump(dataset, (savedir + fname), force=True)
         dl.dump(dataset, os.path.join(savedir, fname), force=True)
 
     def closeEvent(self, event: Any) -> None:
