### 2025-05-14 ec641b1
implements method that stops all running jupyter servers at the end of the session and fixes small issues when saving to pickle
