+++ piva/data_viewer_3d.py
@@ -1894,14 +1894,17 @@ class DataViewer3D(QtWidgets.QMainWindow):
         """
 
         dataset = self.data_set
-
-        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
-            self, 'Select Directory', self.fname[:-len(self.title)]))
-        # savedir = self.fname[:-len(self.title)]
         up = self.util_panel
         file_selection = True
         # Prepare a filename with the .p suffix
         init_fname = '.'.join(self.title.split('.')[:-1] + ['p'])
+        
+        # Choose directory
+        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
+            self, 'Select Directory', self.fname[:-len(self.title)]))
+        # savedir = self.fname[:-len(self.title)]
+        if not savedir:
+            return
 
         while file_selection:
             fname, fname_return_value = QtWidgets.QInputDialog.getText(
