+++ piva/data_viewer_4d.py
@@ -1099,12 +1099,16 @@ class DataViewer4D(QtWidgets.QMainWindow):
 
         # energy axis
         dataset = self.scan
-        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
-            self, 'Select Directory', self.fname[:-len(self.title)]))
-        # savedir = self.fname[:-len(self.title)]
         file_selection = True
         # Prepare a filename with the .p suffix
         init_fname = '.'.join(self.title.split('.')[:-1] + ['p'])
+
+        # Choose directory
+        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
+            self, 'Select Directory', self.fname[:-len(self.title)]))
+        # savedir = self.fname[:-len(self.title)]
+        if not savedir:
+            return
 
         while file_selection:
             fname, fname_return_value = \
