### 2025-05-15 79e9384
improves Testing section and details in installation
