### 2025-05-15 7235c56
adds py312 to CI
