### 2025-06-07 d567467
changes Dataset base class to pytentic.BaseModel
