+++ piva/data_loaders.py
@@ -15,152 +15,160 @@ from warnings import catch_warnings, simplefilter
 from PyQt5.QtWidgets import QDialog, QLabel, QComboBox, QDialogButtonBox, \
     QGridLayout
 from datetime import datetime
-from typing import Union, Any
+from typing import Union, Any, Optional, List, Dict
+from pydantic import BaseModel, Field
 
 
-class Dataset(Namespace):
+class Dataset(BaseModel):
     """
     The generic class inheriting after :class:`~argparse.Namespace` and
     defining structure of the dataset in :mod:`piva`.
     Data files read by corresponding **Dataloader** classes are returned in
     this format.
+
+    Creates a generic `dataset` object inheriting from
+    :class:`~argparse.Namespace` that will be filled with all required data
+    and metadata, depending on what is accessible from files generated by
+    different instruments, and left as :class:`None` otherwise.
+    Asterix indicates attributes that are mandatory for functioning of the
+    **DataViewers**.
+
+    Object contains:
+
+    ===============  ===================  =====================================
+    **attribute**    **type**             **description**
+    **data** *       :class:`np.ndarray`  Acquired data set, always 3D
+                                            matrix (``len(data.shape) = 3``).
+                                            Oriented as: `dim(0)`-scanned
+                                            axis, `dim(1)`- analyzer axis,
+                                            `dim(2)`-energy axis. When scan
+                                            type is a single cut (resulting
+                                            data are 2D), first dimension is
+                                            equal (``data[0, :, :] =
+                                            np.array([0])``)
+    **xscale** *     :class:`np.ndarray`  Axis along the scanned direction,
+                                            units depend on the scan type.
+                                            When scan type is a single cut
+                                            (2D), it is set to
+                                            ``np.array([1])``
+    **yscale** *     :class:`np.ndarray`  Axis along the analyzer slit,
+                                            most likely in [deg]
+    **zscale** *     :class:`np.ndarray`  Axis along the energy direction,
+                                            most likely in [eV]
+    ekin             :class:`np.ndarray`  Energy axis in kinetic energy
+                        | :class:`None`      scale (if default scale is in
+                                            binding energy)
+    kxscale          :class:`np.ndarray`  Momentum axis (saved after
+                        | :class:`None`      conversion) along the scanned
+                                            direction
+    kyscale          :class:`np.ndarray`  Momentum axis (saved after
+                        | :class:`None`      conversion) along the analyzer
+                                            direction
+    x                :class:`float` |     `x` position of the manipulator
+                        :class:`None`
+    y                :class:`float` |     `y` position of the manipulator
+                        :class:`None`
+    z                :class:`float` |     `z` position of the manipulator
+                        :class:`None`
+    theta            :class:`float` |     `theta` angle of the manipulator;
+                        :class:`None`        often referred as `polar`
+
+    phi              :class:`float` |     `phi` angle of the manipulator;
+                        :class:`None`        often referred as `azimuth`
+
+    tilt             :class:`float` |     `tilt` angle of the manipulator
+                        :class:`None`
+    temp             :class:`float` |     Temperature during the experiment
+                        :class:`None`
+    pressure         :class:`float` |     Pressure during the experiment
+                        :class:`None`
+    hv               :class:`float` |     Photon energy used during the
+                        :class:`None`        experiment
+    wf               :class:`float` |     Work function of the analyzer
+                        :class:`None`
+    Ef               :class:`float` |     Correction for the Fermi level
+                        :class:`None`
+    polarization     :class:`str` |       Photon polarization
+                        :class:`None`
[…]
