### 2025-06-07 13a8461
adds pydantic to requirements and updates documentation
