+++ piva/data_loaders.py
@@ -21,13 +21,13 @@ from pydantic import BaseModel, Field
 
 class Dataset(BaseModel):
     """
-    The generic class inheriting after :class:`~argparse.Namespace` and
+    The generic class inheriting after :class:`~pydantic.BaseModel` and
     defining structure of the dataset in :mod:`piva`.
     Data files read by corresponding **Dataloader** classes are returned in
     this format.
 
     Creates a generic `dataset` object inheriting from
-    :class:`~argparse.Namespace` that will be filled with all required data
+    :class:`~pydantic.BaseModel` that will be filled with all required data
     and metadata, depending on what is accessible from files generated by
     different instruments, and left as :class:`None` otherwise.
     Asterix indicates attributes that are mandatory for functioning of the
@@ -219,7 +219,7 @@ class Dataloader:
                       *type* :py:obj:`type`), where *name* stands for how
                       data entry is saved in the file, *label* - how the
                       information should be called in the
-                      :class:`~argparse.Namespace` and *type* - type of the
+                      :class:`~pydantic.BaseModel` and *type* - type of the
                       variable (:py:obj:`float`, :py:obj:`str`, `etc.`)
         :param metadata: if :py:obj:`True`, read only metadata and size of the
                          dataset to display them in :class:`DataBrowser`
@@ -918,7 +918,7 @@ class DataloaderADRESS(Dataloader):
                      format (*name* :py:obj:`str`, *label* :py:obj:`str`,
                      *type* :py:obj:`type`), where *name* stands for how
                      data entry is saved in the file, *label* - how the
-                     information should be called in the :class:`Namespace`
+                     information should be called in the :class:`BaseModel`
                      and *type* - type of the variable (:py:obj:`float`,
                      :py:obj:`str`, `etc.`)
         :param metadata_list: list of strings, usually lines read from loaded
