+++ requirements.txt
@@ -11,6 +11,7 @@ numba>=0.57.0
 matplotlib>=3.3.4
 openpyxl>=3.0.9
 pandas>=1.3.5
+pydantic>=2.0.0
 pyqt5>=5.15.0
 pyqtgraph>=0.13.1
 scipy>=1.6.0
