+++ pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
     "numpy<2.0.0",
     "openpyxl>=3.0.9",
     "pandas>=1.3.5",
+    "pydantic>=2.0.0",
     "pyqt5>=5.15.0",
     "pyqtgraph>=0.13.1",
     "scipy>=1.6.0",
