### 2025-06-07 53eb235
changes Dataloader into abstract base class and updates documentation
