### 2025-06-07 7a62b21
improves implementation of the importer classes and extends documentation
