### 2025-06-07 #231 @pudeIko
Hi all! 👋

Once again, thank you very much for your time, as well as for your valuable and positive reviews!

Over the past few weeks, I’ve been working through the received comments, and I’m happy to say that I’ve addressed the vast majority of them. I believe the package is now ready for another round of review. The only outstanding point is the remark regarding the relatively low test coverage, which I’ll be actively working on in the coming days.

@eigenbrot, below I’ve provided point-by-point responses to your comments. In some cases, I’d like to clarify certain points or explain my reasoning and verify if my understanding is correct. Please let me know if anything remains unclear.

@jsdodge, I noticed that in your review, not all checkboxes were marked as fulfilled. Hopefully, the recent updates resolve those issues, but if anything is still missing or needs correction, I’d kindly ask you to provide a brief description so I can address it accordingly.

Looking forward to your feedback!

---

> Custom Loaders/Plugins
>
> 1. The construction of either a Dataloader subclass or a custom QT widget is generally well described, but how they are imported into PIVA through either DataloaderImporter or PluginImporter is less clear. Examples of these *Importer classes are provided in external files, but I would really like to see a detailed description of the parameters and methods of these classes. They are not mentioned in the otherwise-complete API documentation. A suggestion would be to write abstract bases of these *Importer classes that clearly define the interface.

I was concerned that using abstract base classes (ABCs) might cause issues when `DataloaderImporter` or `PluginImporter` are expected to support multiple custom add-ons. Therefore, I chose a slightly different approach than using ABCs directly. I updated the example to indicate an import method that should not be overridden and is simply called with the implemented `CustomDataloader` as an argument. I also added more detailed in-code documentation and comments, and emphasized in the online documentation that using the provided template is recommended as a tested and reliable solution.
Similarly, for custom widgets, I extended the documentation and added comments in the example code to clarify the exact steps required for correct implementation.

> 2. The fact that a Dataloader subclass must contain a load_data() method is a good indication that Dataloader should probably be an abstract base class.

Absolutely! I have updated the implementation accordingly, along with the corresponding documentation.

> Test System
>
> 1. Per the documentation, PIVA tests are run by calling the test modules directly, which starts a pytest session. This felt weird to me; I think most people are used to simply running pytest in the project root, which does work for PIVA. Running pytest is the method for testing described in CONTRIBUTING.md.

The idea was to allow users to run individual test modules separately, which I thought might be preferable in some cases. Nevertheless, I’ve updated the documentation to present this only as an alternative option, while recommending the standard approach of running pytest from the project root.

> 2. pytest is not listed in the package requirements table. There is also no "test" pip extra.

Both `pytest` and `tox` have now been added as optional dependencies.

> 3. Test coverage for the GUI functionality seems high, but the rest of the package could use more tests. In particular, there don't seem to be any tests of the working_procedures module.

> 4. Despite the presence of a "tox.ini" file, simply running tox locally fails. (I suspect tox is only used for CI builds, so maybe this doesn't matter)

I don’t encounter any issues when running `tox` locally. Could you please share the error message you’re seeing? That would help me identify and resolve the problem more effectively.

> Use of Older Frameworks
>
> There are few places in the code where older methods and frameworks could be replaced with more modern alternatives that would improve the maintainability and robustness of the package. In particular:
>
>
> 1. The Dataset class inherits from python's Namespace class, which allows for properties to be accessing with a nice ds.property syntax. I would highly recommend moving to using either python dataclasses or pydantic BaseModels, both of which are specifically designed for this type of interaction. Pydantic's BaseModel also offers a lot of useful type-validation functionality that may be particularly useful because the internals of the Dataset are tightly constrained to how that class is used by other code.

Changed! Please note that this change breaks compatibility with previously pickled/saved files. Also, `argparse` is still listed as a package requirement. Since it's no longer essential, I created an issue to review remaining occurrences and simplify the code accordingly.

> 2. In a few places (e.g., Notebooks created by the "touch" button in the Data Viewer) file paths are concatenated with a + operator. At the very least I would recommend using os.path.join, or (even better) pathlib.Path objects.

All occurrences should now be corrected. I was already using `os.path` in most places, so I stayed with it for consistency.

> Documentation
>
>
> Overall the documentation is extremely good. My comments are minor:
>
> 1. PIVA's pypi page does not contain a "Documentation" link to the readthedocs site. This is a shame because the documentation is so good!

The link to the documentation has been added to the `pyproject.toml` file and will appear on the PyPI page after the next release.

> 2. I think the "What is piva" section would be more impactful if the multiple strengths/features of piva were given their own subheadings.

Done!

> 3. The "Implemented Dataloaders" table is great, and would be even more great if the individual data loaders were links to the respective class in the API documentation.

Done!

> 4. On the "Data Ingestion and Analysis" page it would be very useful to have an example dataset (the same one used in the example code) so a user can follow along directly. I was eventually able to run the notebook with one of the "test/data" datasets; perhaps elevating one or more of these datasets to an "examples" folder would make their utility more obvious.

The example dataset has been uploaded and is now referenced in the documentation.

> Misc
>
> * Data Viewer line plots show nothing if NaN values are present. The images still render correctly.

Is this something you’d like me to address?

> * astropy is listed as a dependency, but I can't find its usage anywhere

> * Related, the use of astropy.units might be helpful. It's also possible the entire constants module could be replaced with astropy.constants.

I decided to remove `astropy` from the list of dependencies. However, I added a suggestion issue to consider migrating to `astropy.constants` in the future.

> * When Saving data to a pickle (via the "save" button in the Data Viewer) it would be nice to be able to specify the save directory.

Implemented.

> * I found db as the script name to be a little confusing. I kept expecting it to be piva.

Changed! I used to defend it, but many people pointed it out — so I finally gave in. 😄

> * I was expecting the "touch" and "start JL session" buttons to populate a notebook with correct paths for the dataset loaded into the Data Viewer. Instead it looks like they just copied a generic template to the given directory.

I originally tried to implement it that way, but modifying specific lines in notebooks was error-prone and often led to files that couldn’t be executed in Jupyter. So I opted for the more stable template-based approach.

> * It looks like ruff errors are being ignored in CI

I resolved the existing ruff errors and flagged it with `--exit-zero`. This seemed like a good compromise — all issues are reported without blocking CI runs.

> * In tiled window manager, content in windows gets truncated if the window is small enough

I'm not a user of such environments myself, so I'm unsure how to approach this issue. I’m open to suggestions or contributions here.

> * Any opened Jupyter sessions (created with the "start JL session" button) persist after closing all Piva windows.

I’ve implemented a method to stop all opened servers when PIVA is closed.

> * Installation was painless with multiple methods, even those not described in the documentation. FWIW, PIVA appears to work correctly (and the tests pass) for all python versions between 3.10 and 3.13.

> * When clicking the "open in PIT" button the resulting PIT window displays a terminal with the following error in python 3.11 (but not 3.10):

PIT is a separate package developed by a colleague. While I find it incredibly useful — especially for viewing data along arbitrary directions — I’m not responsible for its development or compatibility. I’ve added an info window to make users aware of this.
