+++ piva/tests/viewers_test.py
@@ -4,42 +4,35 @@ Automated test for :class:`~data_viewer_2d.DataViewer2D` and
 most of the functionalities.
 """
 from piva.data_browser import DataBrowser
+from main import db
 from piva.data_viewer_3d import DataViewer3D
 from piva.data_viewer_2d import DataViewer2D
 from piva.fitters import MDCFitter, EDCFitter
 from piva.plot_tool import PlotTool
 from pyqtgraph.Qt.QtCore import Qt
 from PyQt5.QtTest import QTest
+from PyQt5.QtWidgets import QMessageBox
 import numpy as np
 from pkg_resources import resource_filename
 import os
 from piva.working_procedures import get_step
 from typing import Any
+from unittest.mock import patch
 # from piva.utilities_panel import InfoWindow
 
-VTS_MAP = False
 CHECK_3D_Viewer_ = True
 CHECK_2D_Viewer_ = True
 CHECK_EDC_FITTER = True
-CHECK_MDC_FITTER = True
 CHECK_PLOT_TOOL_ = True
 CHECK_LINKING___ = True
-CHECK_K_SPC_CONV = False
+CHECK_K_SPC_CONV = True
+CHECK_MDC_FITTER = True
 
-# EXAMPLE_CUT = pkg_res.resource_filename('piva', 'tests/data/')
 EXAMPLE_CUT = os.path.join(resource_filename('piva', 'tests'), 'data')
-if VTS_MAP:
-    # EXAMPLE_CUT += '/pickle_map.p'
-    EXAMPLE_CUT = os.path.join(EXAMPLE_CUT, 'pickle_map.p')
-    N_SLIDER_E, N_SLIDER_K, N_BINS_E = 80, 160, 3
-    N_E, N_X, N_Y = 20, 3, 9
-    K0_IDX = 374
-else:
-    # EXAMPLE_CUT += '/test_map.p'
-    EXAMPLE_CUT = os.path.join(EXAMPLE_CUT, 'test_map.p')
-    N_SLIDER_E, N_SLIDER_K, N_BINS_E = 20, 26, 3
-    N_E, N_X, N_Y = 20, 14, 3
-    K0_IDX = 83
+EXAMPLE_CUT = os.path.join(EXAMPLE_CUT, 'test_map.p')
+N_SLIDER_E, N_SLIDER_K, N_BINS_E = 20, 26, 3
+N_E, N_X, N_Y = 20, 14, 3
+K0_IDX = 83
 CMAP, CMAP_IDX = 'inferno', 27
 N_GAMMA = 10
 # LONG_WT, SHORT_WT = 700, 5
@@ -62,15 +55,11 @@ class TestViewers:
         browser = DataBrowser()
         qtbot.wait(QTest.qWaitForWindowExposed(browser))
         qtbot.add_widget(browser)
-        qtbot.keyClicks(browser.file_explorer, 'tests')
-        qtbot.keyClick(browser.file_explorer, Qt.Key_Right)
-        qtbot.wait(LONG_WT)
-        qtbot.keyClicks(browser.file_explorer, 'data')
-        qtbot.keyClick(browser.file_explorer, Qt.Key_Right)
+        browser.mb_open_dir(
+            chosen_dir=os.path.join(os.getcwd(), 'piva', 'tests', 'data'))
         qtbot.wait(LONG_WT)
         qtbot.keyClick(browser.file_explorer, Qt.Key_Down)
         qtbot.keyClick(browser.file_explorer, Qt.Key_Down)
-        assert browser.dp_ana_e0.text() == '-'
         qtbot.wait(LONG_WT)
         self.browser = browser
         browser.open_dv(EXAMPLE_CUT)
[…]
