+++ piva/main.py
@@ -8,7 +8,7 @@ from piva.data_browser import DataBrowser
 os.environ['QT_MAC_WANTS_LAYER'] = '1'
 
 
-def db():
+def db(start_event_loop=True):
     version = pkg_resources.require('piva')[0].version
     print( '+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++')
     print( '  PIVA - Photoemission Interface for Visualization and Analysis  ')
@@ -17,6 +17,11 @@ def db():
     app = QApplication(sys.argv)
     DataBrowser()
     sys.exit(app.exec_())
+    # browser = DataBrowser()
+
+    # if start_event_loop:
+    #     sys.exit(app.exec_())
+    # return browser  # allow use without blocking
 
 
 if __name__ == "__main__":
