+++ piva/data_loaders.py
@@ -460,75 +460,75 @@ class Dataloader(ABC):
                     ns.__setattr__('scan_step',
                                    float(tokens[1].split()[0]))
 
-    def _load_ses_pxt(self, filename: str, bl_md: list = None,
-                      metadata: bool = False) -> None:
-        """
-        This is a deprecated, equivalent to original :meth:`load_ses_pxt`
-        method, but based on not maintained anymore :mod:`igor` package.
-        Kept, because old version provided better API, but had dependencies
-        conflicts.
+    # def _load_ses_pxt(self, filename: str, bl_md: list = None,
+    #                   metadata: bool = False) -> None:
+    #     """
+    #     This is a deprecated, equivalent to original :meth:`load_ses_pxt`
+    #     method, but based on not maintained anymore :mod:`igor` package.
+    #     Kept, because old version provided better API, but had dependencies
+    #     conflicts.
 
-        :param filename: absolute path to the file
-        :param bl_md: beamline specific metadata. Not used here, but required
-                      to mach format of other **Dataloaders**. See
-                      :meth:`load_ses_zip` for more info.
-        :param metadata: if :py:obj:`True`, read only metadata and size of the
-                         dataset. See :meth:`load_ses_zip` for more info.
-        """
+    #     :param filename: absolute path to the file
+    #     :param bl_md: beamline specific metadata. Not used here, but required
+    #                   to mach format of other **Dataloaders**. See
+    #                   :meth:`load_ses_zip` for more info.
+    #     :param metadata: if :py:obj:`True`, read only metadata and size of the
+    #                      dataset. See :meth:`load_ses_zip` for more info.
+    #     """
 
-        ds = self.ds
-        pxt = 0#igorpy.load(filename)[0]
+    #     ds = self.ds
+    #     pxt = 0#igorpy.load(filename)[0]
 
-        data = pxt.data.T
-        shape = data.shape
+    #     data = pxt.data.T
+    #     shape = data.shape
 
-        # if file contains single cut
-        if len(shape) == 2:
-            x = 1
-            y = pxt.axis[1].size
-            N_E = pxt.axis[0].size
-            # Make data 3D
-            data = data.reshape((1, y, N_E))
-            # Extract the limits
-            xlims = [1, 1]
-            xscale = start_step_n(*xlims, x)
-            yscale = start_step_n(pxt.axis[1][-1], pxt.axis[1][0], y)
-            zscale = start_step_n(pxt.axis[0][-1], pxt.axis[0][0], N_E)
-        # if it contains multiple cuts, choose which one to load
-        else:
-            multiple_cuts_box = PXT_Dialog(shape[0])
-            box_return_value = multiple_cuts_box.exec()
-            selection = multiple_cuts_box.selection_box.currentIndex()
-            if box_return_value == 0:
-                return
-            data = data[selection, :, :]
-            x = 1
-            y = pxt.axis[1].size
-            N_E = pxt.axis[0].size
-            # Make data 3D
-            data = data.reshape((1, y, N_E))
-            # Extract the limits
-            xlims = [1, 1]
-            xscale = start_step_n(*xlims, x)
-            yscale = start_step_n(pxt.axis[1][-1], pxt.axis[1][0], y)
-            zscale = start_step_n(pxt.axis[0][-1], pxt.axis[0][0], N_E)
+    #     # if file contains single cut
[…]
