+++ .github/workflows/test_with_coverage.yml
@@ -23,8 +23,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-        python-version: ["3.10", "3.11", "3.12"]
+        os: ["ubuntu-latest"]
+        python-version: ["3.10"]
 
     steps:
       - name: Check out repository
