### 2025-06-08 bce08a1
fixes dependencies
