+++ piva/data_viewer_3d.py
@@ -19,6 +19,7 @@ import piva.image_panels as ip
 from piva.utilities_panel import UtilitiesPanel, InfoWindow
 from piva.data_viewer_2d import ORIENTLINES_LINECOLOR
 from piva.cmaps import cmaps
+from piva.config import settings
 if TYPE_CHECKING:
     from piva.data_browser import DataBrowser
 
@@ -1552,19 +1553,20 @@ class DataViewer3D(QtWidgets.QMainWindow):
         ds = self.data_set
         new_dataset = deepcopy(ds)
 
-        # info_box = QMessageBox()
-        # info_box.setIcon(QMessageBox.Information)
-        # info_box.setWindowTitle('K-space conversion.')
-        # msg = 'Conversion might take a while,  ' \
-        #       'depending on the dataset\'s size.\n' \
-        #       'Make sure all parameters are correct and be patient.'.format()
-        # info_box.setText(msg)
-        # info_box.setStandardButtons(QMessageBox.Cancel | QMessageBox.Ok)
-        # choice = info_box.exec()
-        # if choice == QMessageBox.Ok:
-        #     pass
-        # elif choice == QMessageBox.Cancel:
-        #     return
+        if not settings.IS_TESTING:
+            info_box = QMessageBox()
+            info_box.setIcon(QMessageBox.Information)
+            info_box.setWindowTitle('K-space conversion.')
+            msg = 'Conversion might take a while,  ' \
+                'depending on the dataset\'s size.\n' \
+                'Make sure all parameters are correct and be patient.'.format()
+            info_box.setText(msg)
+            info_box.setStandardButtons(QMessageBox.Cancel | QMessageBox.Ok)
+            choice = info_box.exec()
+            if choice == QMessageBox.Ok:
+                pass
+            elif choice == QMessageBox.Cancel:
+                return
 
         if self.util_panel.axes_transform_kz.isChecked():
             ky, _ = wp.hv2kz(ana_axis, scanned_ax, d_scan_ax=d_scan_ax,
@@ -1628,15 +1630,16 @@ class DataViewer3D(QtWidgets.QMainWindow):
             if fail_box.exec() == QMessageBox.Ok:
                 return
 
-        # success_box = QMessageBox()
-        # success_box.setIcon(QMessageBox.Information)
-        # success_box.setWindowTitle('Conversion succeed.')
-        # msg = 'K-space conversion succeed!<br>' \
-        #       'Make sure to save results.'
-        # success_box.setText(msg)
-        # success_box.setStandardButtons(QMessageBox.Ok)
-        # if success_box.exec() == QMessageBox.Ok:
-        #     return
+        if not settings.IS_TESTING:
+            success_box = QMessageBox()
+            success_box.setIcon(QMessageBox.Information)
+            success_box.setWindowTitle('Conversion succeed.')
+            msg = 'K-space conversion succeed!<br>' \
+                'Make sure to save results.'
+            success_box.setText(msg)
+            success_box.setStandardButtons(QMessageBox.Ok)
+            if success_box.exec() == QMessageBox.Ok:
+                return
         return
 
     # def reset_kspace_conversion(self) -> None:
@@ -1901,20 +1904,27 @@ class DataViewer3D(QtWidgets.QMainWindow):
         init_fname = '.'.join(self.title.split('.')[:-1] + ['p'])
         
         # Choose directory
-        savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
-            self, 'Select Directory', self.fname[:-len(self.title)]))
-        # savedir = self.fname[:-len(self.title)]
-        if not savedir:
-            return
+        if settings.IS_TESTING:
+            savedir = os.getcwd()
+            # pass
+        else:
+            savedir = str(QtWidgets.QFileDialog.getExistingDirectory(
[…]
