+++ piva/config.py
@@ -1,0 +1,6 @@
+# Centralized parameters for testing, etc.
+
+class Settings:
+    IS_TESTING = False
+
+settings = Settings()
