+++ tox.ini
@@ -3,10 +3,10 @@
 # shellcheck disable=SC1073
 # shellcheck disable=SC1035
 [tox]
-envlist = py310, py311
+envlist = py310
 
 [testenv]
 deps = -rrequirements.txt
 # execute only specified tests
 commands =
-      pytest piva/tests/dataloaders_test.py piva/tests/viewers_test.py
+      pytest piva/tests/dataloaders_test.py piva/tests/viewers_test.py piva/tests/jupyter_test.py
