+++ piva/data_loaders.py
@@ -12,8 +12,8 @@ import zipfile
 from argparse import Namespace
 from errno import ENOENT
 from warnings import catch_warnings, simplefilter
-from PyQt5.QtWidgets import QDialog, QLabel, QComboBox, QDialogButtonBox, \
-    QGridLayout
+# from PyQt5.QtWidgets import QDialog, QLabel, QComboBox, QDialogButtonBox, \
+#     QGridLayout
 from datetime import datetime
 from typing import Union, Any, Optional, List, Dict
 from pydantic import BaseModel, Field
