+++ piva/working_procedures.py
@@ -9,6 +9,7 @@ from scipy import ndimage
 from scipy.optimize import curve_fit, minimize
 from scipy.signal import convolve2d
 from scipy.fft import fft, ifft, ifftshift
+from piva.data_loaders import Dataset
 
 import piva.constants as const
 
@@ -382,7 +383,6 @@ def get_linear(points: list) -> Callable:
     y = points[1]
 
     pars = np.polyfit(x, y, 1)
-    # fun = lambda arg: pars[0] * arg + pars[1]
 
     def fun(arg):
         return pars[0] * arg + pars[1]
@@ -696,9 +696,9 @@ def get_3D_modulation(qx: np.ndarray, qz: np.ndarray, permutations: list,
     return np.array(res, dtype=complex)
 
 
-def get_self_energy_of_FDW(Q: tuple, K: tuple, ek: Callable,
-                              ek_kwargs: dict, omega: np.ndarray,
-                              Pq: np.ndarray, eta: float = 0.1) -> np.ndarray:
+def get_self_energy_of_FDW(Q: tuple, K: tuple, ek_tb: Callable,
+                           ek_kwargs: dict, omega: np.ndarray,
+                           Pq: np.ndarray, eta: float = 0.1) -> np.ndarray:
     """
     Calculate self-energy of the fluctuating density wave order in a 3D case.
 
@@ -723,19 +723,19 @@ def get_self_energy_of_FDW(Q: tuple, K: tuple, ek: Callable,
     if dim == 1:
         kx = K
         qx = Q
-        res = np.zeros_like(qx.size, dtype=complex)
+        res = np.zeros(qx.size, dtype=complex)
         x_range, y_range, z_range = range(res.shape[0]), range(1), range(1)
     elif dim == 2:
         kx, ky = K
         qx, qy = Q
-        res = np.zeros_like((qx.size, qy.size), dtype=complex)
+        res = np.zeros((qx.size, qy.size), dtype=complex)
         x_range, y_range, z_range = range(res.shape[0]), \
                                     range(res.shape[1]), \
                                     range(1)
     elif dim == 3:
         kx, ky, kz = K
         qx, qy, qz = Q
-        res = np.zeros_like((qx.size, qy.size, qz.size), dtype=complex)
+        res = np.zeros((qx.size, qy.size, qz.size), dtype=complex)
         x_range, y_range, z_range = range(res.shape[0]), \
                                     range(res.shape[1]), \
                                     range(res.shape[2])
@@ -751,17 +751,17 @@ def get_self_energy_of_FDW(Q: tuple, K: tuple, ek: Callable,
                 # doesn't matter
                 if dim == 1:
                     qxi = qx[xi]
-                    ek = ek(kx + qxi, **ek_kwargs)
+                    ek = ek_tb(kx + qxi, **ek_kwargs)
                     ek = np.array(np.where(ek == 0, 1e-10, ek), dtype=complex)
                     res = np.sum(Pq / ((omega + eta * 1.j) - ek))
                 elif dim == 2:
                     qxi, qyi = qx[xi], qy[yi]
-                    ek = ek(kx + qxi, ky=kx + qyi, **ek_kwargs)
+                    ek = ek_tb(kx + qxi, ky=kx + qyi, **ek_kwargs)
                     ek = np.array(np.where(ek == 0, 1e-10, ek), dtype=complex)
                     res[xi, yi] = np.sum(Pq / ((omega + eta * 1.j) - ek))
                 elif dim == 3:
                     qxi, qyi, qzi = qx[xi], qy[yi], qz[zi]
[…]
