+++ piva/data_viewer_2d.py
@@ -4,9 +4,9 @@ from copy import deepcopy
 from typing import TYPE_CHECKING, Any
 import numpy as np
 from pyqtgraph.Qt import QtWidgets
-from pyqtgraph.Qt.QtCore import Qt
+# from pyqtgraph.Qt.QtCore import Qt
 from PyQt5.QtWidgets import QMessageBox
-from PyQt5.QtCore import QTimer
+# from PyQt5.QtCore import QTimer
 
 import piva.working_procedures as wp
 import piva.data_loaders as dl
