+++ .github/workflows/test_with_coverage.yml
@@ -8,7 +8,6 @@ on:
   push:
     branches:
       - main
-      - pyOS-review
 
 defaults:
   run:
@@ -23,8 +22,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest"]
-        python-version: ["3.10"]
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        python-version: ["3.10", "3.11", "3.12"]
 
     steps:
       - name: Check out repository
