### 2025-06-13 #231 @eigenbrot
## Package Review 2

#### General

Thanks, @pudeIko, for your changes, I can tell a lot of hard work went into these updates!
My new comments are below. Anything from before that I don't mention is something that I have no additional comments on; thank your for addressing and/or clarifying those issues.

#### `*Importer` Classes

The additional documentation and examples for the `*Importer` classes really help a lot. Nice work!
Regarding your comment about making `DataloaderImporter` an ABC, I understand what you're saying about multiple custom
loaders and I think a solution could be something like this (built on your example)

```python
from abc import ABC, abstractmethod

class DataloaderImporterBase(ABC):
    """
    Base Dataloader importer for custom **Dataloaders** written for PIVA.

    Add any custom `DataLoader` subclasses to the `data_loaders` property.
    """

    def __init__(self, data_browser: DataBrowser) -> None:
        """
        Initialize DataloaderImporter and export custom written loaders to the
        record of available **Dataloaders**.

        :param data_browser: `DataBrowser` of the current session
        """

        # assign a reference to the DataBrowser
        self.db = data_browser
        self.db.dp_dl_picker.insertSeparator(self.db.dp_dl_picker.count())

        # use this method to add the loader to the DataBrowser's record!
        for loader in self.custom_loader_list:
            self.add_dataloader(loader)

    @property
    @abstractmethod
    def custom_loader_list(self) -> list[DataLoader]:
        """The list of custom `DataLoaders` to add to PIVA."""
        pass

    @final
    def add_dataloader(self, loader: Dataloader) -> None:
        """
        Method for exporting implemented custom Dataloader to the
        DataBrowser's list of available dataloaders.

        Should not be overridden!

        :param loader: the implemented CustomDataloader that needs to be
               imported
        """

        dl_label = '{}'.format(loader.name)

        self.db.dp_dl_picker.addItem(dl_label)
        self.db.add_dataloader_to_record(dl_label, loader)

        print('\t', dl_label)
```

and then a user's "piva_dataloader_importer.py" could look like

```python
from piva.somewhere import DataloaderImporterBase
from .my_loaders import Loader1, Loader2

class MyGreatImporter(DataloaderImporterBase):
    @property
    def custom_loader_list(self) -> list[DataLoader]:
        return [Loader1, Loader2]
```
As I see it, the primary advantage of this method would be that `DataloaderImporterBase` would be a first-class object in
the repo and therefore included in the excellent API documentation. A similar treatment could be given to the `PluginImporter`, as well.

To be clear, this is just a suggestion; I think the additions to the documentation and examples are sufficient to call this issue resolved.

#### Tests
Thank you for adding test coverage, specifically of the `working_procedures` module. Thank you also for adding more detail
to the online documentation about running tests. On this note, I think CONTRIBUTING.md may now be a bit out of date; perhaps the
"Install Dependencies" step should have `pip install -e .[test]` instead of using the requirements file?

I still have two concerns about the tests:

1. It looks like the viewer, jupyter, and working procedures tests are only run on `ubuntu-latest` with python 3.10 in the github CI. Is there a reason to skip other OS's/versions?
2. I can't run the tests for the `working_procedures` module on my computer. When I try to run them they take up 100% of a CPU for about 20 minutes before the processes is killed by my kernel for using up all (~50 GB) of my RAM. On Github I see that these tests run in ~15s, so I'm not sure what's going on here.

Related, I no longer get an error when running `tox`, but the tox config only seems to run the `working_procedures` test.
Furthermore, it looks like `tox` isn't used in Github actions. If this is true, it might be better to just remove `tox` altogether.

#### New `Dataset` class
Thank you for using `BaseModel` as the base for the `Dataset` class. I just have 2 very minor notes on that implementation:

1. The required attributes `data`, `xscale`, `yscale`, and `zscale` would be better defined with no default (right now they default to `None`). This would cause `pydantic` to raise a validation error if they are not provided.
2. IIRC the type-hint `SOME_TYPE | None` has been preferred over `Optional[SOME_TYPE]` for a while (since python 3.10, maybe?).

e.g.
```python
class Dataset(BaseModel):
    ...

    data: np.ndarray
    xscale: np.ndarray
    yscale: np.ndarray
    zscale: np.ndarray
    ekin: np.ndarray | None = None
    ...
```

#### CI Linting Step
I'm not sure I see the point of using `--exit-zero` with `ruff` in the linting CI step. This feels barely different than simply ignoring the step; in both cases any linting errors would be ignored in an automated run and the onus is on a human to manually fix things. In other words, I'd argue that the only way to *ensure* correct code formatting is to have the CI break if the formatting is bad. I don't know how stringint PyOpenSci is about this requirement, so I'll leave it up to @crhea93  for the final say on whether this would block approval.

If you decide to not ignore `ruff` errors in the CI then I would suggest looking into using [ruff's pre-commit hook](https://github.com/astral-sh/ruff-pre-commit) to automatically fix any formatting errors everytime a commit is made.

#### Misc

* The "Citing" section of README.md still says "TBD".
* The requirements.txt and the "Dependencies" section of the online docs are out of date w.r.t to the removal of `astropy`.
* Regarding NaN values causing the line plots to disappear: I'm fine with whatever you decide. Maybe in the field of ARPES data NaNs are uncommon. They occur often in data I look at (astronomy). If it was something simple like plotting `(x[~np.isnan(y)], y[~np.isnan(y)])` instead of just `(x, y)` then it may be worth it. Either way is fine with me.
* The online docs still make reference to the command name `db` on the "Data Browser" page and in the note box regarding wayland on the "Installation" page.
* I'm having an issue with the new "Au-test_spectrum.p" file. If I open it in a notebook as suggested on the "Data Ingestion and Analysis" page then the `dataset` objet has no `.data` attribute. If I try to open the same file with `piva` itself I get the following error in the consol window:
```
Traceback (most recent call last):
  File "~/piva/piva/data_browser.py", line 118, in launch_piva
    self.open_dv(fname)
  File "~/piva/piva/data_browser.py", line 194, in open_dv
    raise e
  File "~/piva/piva/data_browser.py", line 182, in open_dv
    if data_set.xscale.size == 1:
AttributeError: 'NoneType' object has no attribute 'size'
```
* Nice work with the new Jupter shutdown capabilities!
