### 2025-06-13 #231 @crhea93
Thank you for this @eigenbrot ! We encourage following standard lint practices but understand that it isn't always feasible, so this would not block approval :)

@pudeIko If you have the bandwidth you can look at the suggestions from @eigenbrot in his comment. That being said, don't worry too much about this!
