### 2025-06-19 d0eae28
ruff format and changes from the 2nd review
