+++ piva/utilities_panel.py
@@ -5,11 +5,25 @@ from sys import platform
 from datetime import datetime
 from typing import Union, Any, TYPE_CHECKING
 import numpy as np
-from PyQt5.QtWidgets import QWidget, QLabel, QCheckBox, QComboBox, \
-    QDoubleSpinBox, QSpinBox, QPushButton, QLineEdit, QMainWindow, \
-    QDialogButtonBox, QMessageBox, QScrollArea, QTableWidget, QVBoxLayout
+from PyQt5.QtWidgets import (
+    QWidget,
+    QLabel,
+    QCheckBox,
+    QComboBox,
+    QDoubleSpinBox,
+    QSpinBox,
+    QPushButton,
+    QLineEdit,
+    QMainWindow,
+    QDialogButtonBox,
+    QMessageBox,
+    QScrollArea,
+    QTableWidget,
+    QVBoxLayout,
+)
 from PyQt5.QtWidgets import QTableWidgetItem as QTabItem, QSizePolicy
 from pyqtgraph.Qt import QtCore, QtGui, QtWidgets
+
 if TYPE_CHECKING:
     from piva.data_viewer_2d import DataViewer2D
     from piva.data_viewer_3d import DataViewer3D
@@ -34,7 +48,7 @@ QTabWidget{background-color: rgb(64, 64, 64);}
 """
 SIGNALS = 5
 MY_CMAPS = True
-DEFAULT_CMAP = 'coolwarm'
+DEFAULT_CMAP = "coolwarm"
 
 bold_font = QtGui.QFont()
 bold_font.setBold(True)
@@ -46,8 +60,12 @@ class UtilitiesPanel(QWidget):
     visualization and basic analysis utilities.
     """
 
-    def __init__(self, main_window: Union[DataViewer2D, DataViewer3D],
-                 name: str = None, dim: int = 3) -> None:
+    def __init__(
+        self,
+        main_window: Union[DataViewer2D, DataViewer3D],
+        name: str = None,
+        dim: int = 3,
+    ) -> None:
         """
         Initialize Utilities panel window
 
@@ -66,30 +84,30 @@ class UtilitiesPanel(QWidget):
         self.dim = dim
         self.jl_session_running = False
 
-        self.close_button = QPushButton('close')
-        self.save_button = QPushButton('save')
-        self.hide_button = QPushButton('hide tabs')
-        self.pit_button = QPushButton('open in PIT')
+        self.close_button = QPushButton("close")
+        self.save_button = QPushButton("save")
+        self.hide_button = QPushButton("hide tabs")
+        self.pit_button = QPushButton("open in PIT")
 
         self.buttons = QWidget()
         self.buttons_layout = QtWidgets.QGridLayout()
-        self.buttons_layout.addWidget(self.close_button,    1, 0)
-        self.buttons_layout.addWidget(self.save_button,     2, 0)
+        self.buttons_layout.addWidget(self.close_button, 1, 0)
+        self.buttons_layout.addWidget(self.save_button, 2, 0)
         if self.dim in (2, 4):
-            self.file_mdc_fitter_button = QPushButton('MDC fitter')
-            self.file_edc_fitter_button = QPushButton('EDC fitter')
+            self.file_mdc_fitter_button = QPushButton("MDC fitter")
+            self.file_edc_fitter_button = QPushButton("EDC fitter")
             self.buttons_layout.addWidget(self.file_mdc_fitter_button, 3, 0)
             self.buttons_layout.addWidget(self.file_edc_fitter_button, 4, 0)
         if self.dim == 2:
-            self.buttons_layout.addWidget(self.pit_button,      5, 0)
+            self.buttons_layout.addWidget(self.pit_button, 5, 0)
         elif self.dim == 3:
[…]
