+++ docs/misc/piva_plugin_importer.py
@@ -15,24 +15,24 @@ class PluginImporter:
         :param data_browser: `DataBrowser` of the current session
         """
 
-        # create a reference to the DataBrowser and a "User plugins" menu 
+        # create a reference to the DataBrowser and a "User plugins" menu
         self.db = data_browser
-        self.db.user_menu = self.db.menu_bar.addMenu('&User plugins')
+        self.db.user_menu = self.db.menu_bar.addMenu("&User plugins")
 
         self.import_widget1()
 
     def import_widget1(self) -> None:
         """
         Example method for importing and setting up custom widget.
-        
-        The example creates a menu action (and a shortcut) that triggers 
-        the opening of the custom widget (i.e., executes the method that 
+
+        The example creates a menu action (and a shortcut) that triggers
+        the opening of the custom widget (i.e., executes the method that
         initializes it), and adds it to the DataBrowser.
         """
 
-        open_widget = QAction('Custom widget', self.db)
-        open_widget.setStatusTip('Custom widget')
-        open_widget.setShortcut('Ctrl+D')
+        open_widget = QAction("Custom widget", self.db)
+        open_widget.setStatusTip("Custom widget")
+        open_widget.setShortcut("Ctrl+D")
         open_widget.triggered.connect(self.open_widget1)
         self.db.user_menu.addAction(open_widget)
 
@@ -42,8 +42,6 @@ class PluginImporter:
         :class:`QAction` assigned to the **CustomWidget** is triggered.
         """
 
-        idx = 'tmp'
-        self.db.custom_plugins[idx] = \
-            cw1.CustomWidget(data_browser=self.db, index=idx)
-        print('\t', 'CustomWidget')
-
+        idx = "tmp"
+        self.db.custom_plugins[idx] = cw1.CustomWidget(data_browser=self.db, index=idx)
+        print("\t", "CustomWidget")
