+++ piva/data_viewer_4d.py
@@ -21,7 +21,7 @@ app_style = """
 QMainWindow{background-color: rgb(64,64,64);}
 QTabWidget{background-color: rgb(64,64,64);}
 """
-DEFAULT_CMAP = 'viridis'
+DEFAULT_CMAP = "viridis"
 ORIENTLINES_LINECOLOR = (164, 37, 22, 255)
 
 
@@ -78,7 +78,7 @@ class DataHandler4D:
         """
 
         # self.data = ip.TracedVariable(data, name='data')
-        self.data = ip.CustomTracedVariable(data, name='data')
+        self.data = ip.CustomTracedVariable(data, name="data")
         self.axes = np.array(axes, dtype="object")
         self.norm_data = wp.normalize(self.data.get_value()[0, :, :].T).T
 
@@ -115,8 +115,7 @@ class DataHandler4D:
         x_ax, y_ax = [], []
         for xi in range(nx):
             for yi in range(ny):
-                self.scan_data[xi, yi] = \
-                    self.main_window.scan[xi, yi].data[0, :, :]
+                self.scan_data[xi, yi] = self.main_window.scan[xi, yi].data[0, :, :]
                 x_ax.append(self.main_window.scan[xi, yi].x)
                 y_ax.append(self.main_window.scan[xi, yi].y)
         self.x_ax = np.unique(np.array(x_ax))
@@ -141,8 +140,8 @@ class DataHandler4D:
                 if up.image_normalize.isChecked():
                     norm_along = up.image_normalize_along.currentIndex()
                     tmp_data = wp.normalize(
-                        self.main_window.scan[xi, yi].data[0, :, :],
-                        axis=norm_along)
+                        self.main_window.scan[xi, yi].data[0, :, :], axis=norm_along
+                    )
                 else:
                     tmp_data = self.main_window.scan[xi, yi].data[0, :, :]
                 self.scan_data[xi, yi] = tmp_data
@@ -161,19 +160,17 @@ class DataHandler4D:
         tmp_scan_data = np.zeros((nx, ny, ne, nk))
         for xi in range(nx):
             for yi in range(ny):
-                tmp_scan_data[xi, yi] = \
-                    self.main_window.scan[xi, yi].data[0, :, :]
+                tmp_scan_data[xi, yi] = self.main_window.scan[xi, yi].data[0, :, :]
         if opt == 0:
             self.raster_data = np.sum(np.sum(tmp_scan_data, axis=2), axis=2)
         elif opt == 1:
             for xi in range(tmp_scan_data.shape[0]):
                 for yi in range(tmp_scan_data.shape[1]):
-                    flat = tmp_scan_data[xi, yi].flatten() + 1.
+                    flat = tmp_scan_data[xi, yi].flatten() + 1.0
                     cut_off = int(flat.size * 0.05)
                     low = np.sort(flat)[:cut_off]
                     high = np.sort(flat)[-cut_off:]
-                    self.raster_data[xi, yi] = \
-                        np.average(high) / np.average(low)
+                    self.raster_data[xi, yi] = np.average(high) / np.average(low)
         elif opt == 2:
             for xi in range(tmp_scan_data.shape[0]):
                 for yi in range(tmp_scan_data.shape[1]):
@@ -187,8 +184,13 @@ class DataViewer4D(QtWidgets.QMainWindow):
     Main window of the 4D data.
     """
 
-    def __init__(self, data_browser: DataBrowser, data_set: dl.Dataset,
-                 index: str = None, slice: bool = False) -> None:
+    def __init__(
+        self,
+        data_browser: DataBrowser,
+        data_set: dl.Dataset,
+        index: str = None,
[…]
