+++ piva/config.py
@@ -1,6 +1,8 @@
 # Centralized parameters for testing, etc.
 
+
 class Settings:
     IS_TESTING = False
+
 
 settings = Settings()

[…]
