### 2025-06-19 e6fe66e
run full test coverage on all OSs
