+++ .github/workflows/test_with_coverage.yml
@@ -65,33 +65,33 @@ jobs:
           --cov-config=.coveragerc piva/tests/dataloaders_test.py
 
       - name: Test viewers - install prerequisites
-        if: ${{matrix.os=='ubuntu-latest' && matrix.python-version=='3.10'}}
+        if: ${{matrix.python-version=='3.10'}}
         run: >
           sudo apt install libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0
           libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0
           libxcb-xfixes0 x11-utils
 
       - name: Test viewers - configure
-        if: ${{matrix.os=='ubuntu-latest' && matrix.python-version=='3.10'}}
+        if: ${{matrix.python-version=='3.10'}}
         run: >
           /sbin/start-stop-daemon --start --quiet --pidfile
           /tmp/custom_xvfb_99.pid --make-pidfile --background --exec
           /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
 
       - name: Test viewers - run tests
-        if: ${{matrix.os=='ubuntu-latest' && matrix.python-version=='3.10'}}
+        if: ${{matrix.python-version=='3.10'}}
         run: >
           pytest --cov=$GITHUB_WORKSPACE --cov-append --cov-report=xml
           --cov-config=.coveragerc piva/tests/viewers_test.py
 
       - name: Test jupyter functionalities
-        if: ${{matrix.os=='ubuntu-latest' && matrix.python-version=='3.10'}}
+        if: ${{matrix.python-version=='3.10'}}
         run: >
           pytest --cov=$GITHUB_WORKSPACE --cov-append --cov-report=xml
           --cov-config=.coveragerc piva/tests/jupyter_test.py
 
       - name: Test working_procedures module
-        if: ${{matrix.os=='ubuntu-latest' && matrix.python-version=='3.10'}}
+        if: ${{matrix.python-version=='3.10'}}
         run: >
           pytest --cov=$GITHUB_WORKSPACE --cov-append --cov-report=xml
           --cov-config=.coveragerc piva/tests/wp_test.py
