### 2025-06-20 255dec9
extend test matrix
