### 2025-06-20 #231 @pudeIko
Thanks again for the detailed remarks, @eigenbrot, and for pointing out the outdated parts of the documentation!

As before, here’s my point-by-point response and I'm looking forward to your feedback!

---

> Regarding your comment about making DataloaderImporter an ABC, I understand what you're saying about multiple custom loaders and I think a solution could be something like this (built on your example)
>
>     from abc import ABC, abstractmethod
>
>
>     class DataloaderImporterBase(ABC):
>         """
>         Base Dataloader importer for custom **Dataloaders** written for PIVA.
>
>         Add any custom `DataLoader` subclasses to the `data_loaders` property.
>         """
>

I really liked the approach and gave it a try, but I encountered the following error:

```
Traceback (most recent call last):
  File "/Users/kaskadermike/opt/anaconda3/envs/piva/bin/db", line 33, in <module>
    sys.exit(load_entry_point('piva', 'console_scripts', 'db')())
  File "/Users/kaskadermike/git/piva/piva/main.py", line 18, in db
    DataBrowser()
  File "/Users/kaskadermike/git/piva/piva/data_browser.py", line 70, in __init__
    self.load_custom_data_loaders()
  File "/Users/kaskadermike/git/piva/piva/data_browser.py", line 850, in load_custom_data_loaders
    self.dli = getattr(dli, "DataloaderImporterBase")(self)
TypeError: Can't instantiate abstract class DataloaderImporterBase with abstract method custom_loader_list

```
I believe this is similar to the issue with the Datasets objects (related to initialization and validation, described below). Since the DataBrowser searches for a specific DataloaderImported module and executes code from it, switching to an ABC would require reworking the entire importing mechanism. So, if the previous changes resolve the issue, I’d prefer to stick with the current approach.

> Tests
> Thank you for adding test coverage, specifically of the working_procedures module.
>
> Thank you also for adding more detail to the online documentation about running tests. On this note, I think CONTRIBUTING.md may now be a bit out of date; perhaps the "Install Dependencies" step should have pip install -e .[test] instead of using the requirements file?

Done!

> I still have two concerns about the tests:
> 1.	It looks like the viewer, jupyter, and working procedures tests are only run on ubuntu-latest with python 3.10 in the github CI. Is there a reason to skip other OS's/versions?
>

You're right — the tests for the viewer, Jupyter, and working procedures currently run only on `ubuntu-latest` with Python 3.10 in the GitHub CI. The main reason for this is that I had significant trouble getting them to work on other OSes, I believe due to the use of qtbot and GUI components. After quite a bit of research, I managed to find a working setup for Ubuntu, but on macOS and Windows the tests consistently fail during `viewers_test`.

That said, I have extended the coverage to include the remaining Python versions — just limited to Ubuntu for now.

As for the tests that do pass across all platforms: the `dataloaders` and `working_procedures` tests. I'm running the former because it effectively verifies that the package and its core entry points install and initialize correctly. The latter I chose to exclude purely out of thrift — it takes about 10 minutes to complete on macOS, and considering GitHub’s time multipliers, it could easily consume all of my available free CI minutes.

> 2.	I can't run the tests for the working_procedures module on my computer. When I try to run them they take up 100% of a CPU for about 20 minutes before the processes is killed by my kernel for using up all (~50 GB) of my RAM. On Github I see that these tests run in ~15s, so I'm not sure what's going on here.

I encountered the same problem and traced it back to a few of the fitting routines. The issue seemed to be caused by the size and complexity of the data being fitted. I resolved it by downsampling the dataset used in the tests, which allowed them to run successfully both locally and in GitHub Actions.

That said, I'm not entirely sure why it's still not completing on your machine, but my guess is the issue is somewhere there.

> Related, I no longer get an error when running tox, but the tox config only seems to run the working_procedures test. Furthermore, it looks like tox isn't used in Github actions. If this is true, it might be better to just remove tox altogether.

My mistake! It was only running `working_procedures_test` because that’s what I had tested last. Anyways, following your advice, I’ve now moved `tox.ini` to `.gitignore`.

> New Dataset class
>
> Thank you for using BaseModel as the base for the Dataset class. I just have 2 very minor notes on that implementation:
> 1.	The required attributes data, xscale, yscale, and zscale would be better defined with no default (right now they default to None). This would cause pydantic to raise a validation error if they are not provided.
>

Done! However, since these attributes are populated later, I was initially running into a `ValidationError`. To handle this, I'm now using `Dataset.model_construct()` during initialization and validating the instance upon return. This should effectively achieve the same goal. I’ve also updated the custom examples accordingly.

> 2.	IIRC the type-hint SOME_TYPE | None has been preferred over Optional[SOME_TYPE] for a while (since python 3.10, maybe?).
> e.g.
>
> class Dataset(BaseModel):
>     ...
>
>     data: np.ndarray
>     xscale: np.ndarray
>     yscale: np.ndarray
>     zscale: np.ndarray
>     ekin: np.ndarray | None = None
>     ...
>

Done!

> CI Linting Step
> I'm not sure I see the point of using --exit-zero with ruff in the linting CI step. This feels barely different than simply ignoring the step; in both cases any linting errors would be ignored in an automated run and the onus is on a human to manually fix things. In other words, I'd argue that the only way to ensure correct code formatting is to have the CI break if the formatting is bad. I don't know how stringint PyOpenSci is about this requirement, so I'll leave it up to [@crhea93](https://github.com/crhea93) for the final say on whether this would block approval. If you decide to not ignore ruff errors in the CI then I would suggest looking into using [ruff's pre-commit hook](https://github.com/astral-sh/ruff-pre-commit) to automatically fix any formatting errors everytime a commit is made.
>

I completely agree — it felt the same to me, but I came across it as a commonly suggested workaround. The issue was that, for some reason, `ruff format` was throwing errors in CI, even though everything worked fine locally. That said, the number of corrections was small, so I applied them manually based on the GitHub output, and now it runs cleanly!

> Misc
>
> •	The "Citing" section of README.md still says "TBD".

I’ve left it as "TBD" for now, as I’m planning to update it with the full reference once the JOSS publication is out.

> •	The requirements.txt and the "Dependencies" section of the online docs are out of date w.r.t to the removal of astropy.

Done!

> •	Regarding NaN values causing the line plots to disappear: I'm fine with whatever you decide. Maybe in the field of ARPES data NaNs are uncommon. They occur often in data I look at (astronomy). If it was something simple like plotting (x[~np.isnan(y)], y[~np.isnan(y)]) instead of just (x, y) then it may be worth it. Either way is fine with me.

I hadn’t really considered NaNs before — which probably means they’re quite rare in my case 😄. In ARPES, they typically only occur when an entire scan fails.
I implemented your suggestion by filtering out the NaNs in the viewers, and everything seems to work fine with my data and during testing. Let me know if it improves the behavior on your end. I applied the change only to the viewers; since the fitters are very ARPES-specific, I left those unchanged.

> •	The online docs still make reference to the command name db on the "Data Browser" page and in the note box regarding wayland on the "Installation" page.

Done!

> •	I'm having an issue with the new "Au-test_spectrum.p" file. If I open it in a notebook as suggested on the "Data Ingestion and Analysis" page then the dataset objet has no .data attribute. If I try to open the same file with piva itself I get the following error in the consol window:
>
> Traceback (most recent call last):
>   File "~/piva/piva/data_browser.py", line 118, in launch_piva
>     self.open_dv(fname)
>   File "~/piva/piva/data_browser.py", line 194, in open_dv
>     raise e
>   File "~/piva/piva/data_browser.py", line 182, in open_dv
>     if data_set.xscale.size == 1:
> AttributeError: 'NoneType' object has no attribute 'size'

Yes, that issue was caused by a conflict with older files saved using the deprecated `Dataset(Namespace)` class. It should be resolved now!

> •	Nice work with the new Jupter shutdown capabilities!

Thanks!
