### 2025-06-20 #231 @eigenbrot
Thanks, @pudeIko. I probably won't be able to get to this until early next week, but, judging from your responses, it sounds like the my next look will be pretty quick!
