### 2025-06-24 #231 @eigenbrot
## Package Review 3

Thanks again, @pudeIko! These new updates are great and I think I'm ready to approve this review.
I have a few minor comments that you can implement if you want; nothing that will block acceptance.

* I see your point as to why the ABC doesn't work in `DataloaderImporter` and agree that moving in that direction would require a full rewrite of the import mechanisms. Thanks for indulging me, anyway!
* Thank you for extending the test matrix to include more python versions.
* I hear what you're saying about the length of the `working_procedures` test. I poked into this a bit and (on my machine, at least) the `popt, cov = curve_fit(wp.two_lorentzians, peak_x, peak_y)` line is what is taking all of the time and RAM. If I comment out that line then the all the tests run in a second or two. I don't know the details of these tests so I'm not sure if that's useful or not. I think at this point I'm fine with the situation where the tests at least run on the Github CI. You might consider leaving a note somewhere that the `working_procedures` tests are known to hang up on some systems.
* I see the issue you ran into with the `Dataset` base model, and I think the `.model_construct()` solution is great. I didn't even know about that method!
* Related to the above: given that you are essentially instantiating an empty `Dataset` and then filling in the relevant attributes, you might consider setting the [validate_assignment](https://docs.pydantic.dev/latest/api/config/#pydantic.config.ConfigDict.validate_assignment) option to `True` in the `config_dict` of `Dataset`. This will ensure that all attributes are *always* the correct type.
* I'm glad you were able to get the linting sorted. Having an automatic linter is the path to formatting bliss :grin:
* Thanks for fixing the NaN value issue in the viewers. My data look great now :grinning:
* Thanks for fixing the issue with the ""Au-test_spectrum.p" test file. It now loads correctly and I can replicate the results shown in the "Data Ingestion and Analysis" page.

@crhea93, unless @pudeIko has any other comments/changes, I'm ready to sign off on this review. How do I do that?
