### 2025-06-25 8e664aa
add logos in badges
