### 2025-06-25 5188677
move standard return actions into Dataloader base class
