+++ piva/data_loaders.py
@@ -593,9 +593,27 @@ class Dataloader(ABC):
         # print(isinstance(scan, np.ndarray))
         return scan
 
-    def validate_at_return(self):
-        self.ds.validate_assignment = True
-        return Dataset.model_validate(self.ds.model_dump())
+    def validate_at_return(self, filename: str):
+        """
+        Validate that the Dataset was correctly populated with data and add
+        original file information to the data provenance record.
+
+        :param filename: absolute path to the file
+        :return: loaded dataset with available metadata
+        """
+
+        if self.raster:
+            for xi in range(self.scan.shape[0]):
+                for yi in range(self.scan.shape[1]):
+                    dsi = self.scan[xi, yi]
+                    dsi.add_org_file_entry(filename, self.name)
+                    dsi.validate_assignment = True
+                    self.scan[xi, yi] = Dataset.model_validate(dsi.model_dump())
+            return self.scan
+        else:
+            self.ds.add_org_file_entry(filename, self.name)
+            self.ds.validate_assignment = True
+            return Dataset.model_validate(self.ds.model_dump())
 
 
 class DataloaderPickle(Dataloader):
@@ -641,14 +659,7 @@ class DataloaderPickle(Dataloader):
                 except Exception as e:
                     raise e
 
-        if self.raster:
-            for xi in range(self.scan.shape[0]):
-                for yi in range(self.scan.shape[1]):
-                    self.scan[xi, yi].add_org_file_entry(filename, self.name)
-            return self.scan
-        else:
-            self.ds.add_org_file_entry(filename, self.name)
-            return self.validate_at_return()
+        return self.validate_at_return(filename)
 
 
 class DataloaderSIS(Dataloader):
@@ -685,14 +696,7 @@ class DataloaderSIS(Dataloader):
         else:
             raise NotImplementedError
 
-        if self.raster:
-            for xi in range(self.scan.shape[0]):
-                for yi in range(self.scan.shape[1]):
-                    self.scan[xi, yi].add_org_file_entry(filename, self.name)
-            return self.scan
-        else:
-            self.ds.add_org_file_entry(filename, self.name)
-            return self.validate_at_return()
+        return self.validate_at_return(filename)
 
     def load_h5(self, filename: str, metadata: bool = False) -> None:
         """
@@ -859,8 +863,7 @@ class DataloaderADRESS(Dataloader):
         else:
             raise NotImplementedError
 
-        self.ds.add_org_file_entry(filename, self.name)
-        return self.validate_at_return()
+        return self.validate_at_return(filename)
 
     def load_h5(self, filename: str, metadata: bool = False) -> None:
         """
@@ -1030,14 +1033,7 @@ class DataloaderBloch(Dataloader):
         else:
             raise NotImplementedError
 
-        if self.raster:
-            for xi in range(self.scan.shape[0]):
-                for yi in range(self.scan.shape[1]):
-                    self.scan[xi, yi].add_org_file_entry(filename, self.name)
-            return self.scan
-        else:
-            self.ds.add_org_file_entry(filename, self.name)
-            return self.validate_at_return()
+        return self.validate_at_return(filename)
 
 
 class DataloaderI05(Dataloader):
@@ -1051,11 +1047,11 @@ class DataloaderI05(Dataloader):
     def __init__(self):
         super(DataloaderI05, self).__init__()
 
-    def load_data(self, fname: str, metadata: bool = False) -> Dataset:
+    def load_data(self, filename: str, metadata: bool = False) -> Dataset:
         """
         Recognize correct format and load data from the file.
 
-        :param fname: absolute path to the file
+        :param filename: absolute path to the file
[…]
