### 2025-06-25 #231 @pudeIko
Thanks a lot @eigenbrot for your prompt response!

Answering two of your comments:

* Since it's just this one problematic function, I decided to comment it out for now — it's not particularly relevant and doesn't significantly impact test coverage.
* I experimented with `validate_assignment=True`, but ran into errors again. However, I think I found a good workaround by adding a method to the base class:

```python
class Dataloader(ABC):

...
      def validate_at_return(self, filename: str):
          """
          Validate that the Dataset was correctly populated with data and add
          original file information to the data provenance record.

          :param filename: absolute path to the file
          :return: loaded dataset with available metadata
          """

          if self.raster:
              for xi in range(self.scan.shape[0]):
                  for yi in range(self.scan.shape[1]):
                      dsi = self.scan[xi, yi]
                      dsi.add_org_file_entry(filename, self.name)
                      dsi.validate_assignment = True
                      self.scan[xi, yi] = Dataset.model_validate(dsi.model_dump())
              return self.scan
          else:
              self.ds.add_org_file_entry(filename, self.name)
              self.ds.validate_assignment = True
              return Dataset.model_validate(self.ds.model_dump())
```

Then I adjusted all dataloaders to use this method when returning:

```python
class DataloaderPickle(Dataloader):

    def load_data(self, filename: str, metadata: bool = False) -> Dataset:
        ...

        return self.validate_at_return()
```

---

I'm really happy to hear that the package has passed acceptance — and thank you so much for all the remarks and suggestions! I truly appreciate the improvements they led to!
