### 2025-06-26 #231 @crhea93
@eigenbrot Thank you very much for this hard work (and thank you @pudeIko for the detailed responses). I'll make a note that you (@eigenbrot ) accept the package as ready for publication in the PyOpenSci ecosystem. At this point, you have no more duties as a reviewer :D

Once @jsdodge wraps up his review (this is not me trying to rush you!!), I'll  begin the procedure of a formal acceptance :)
