### 2025-07-30 #231 @pudeIko
Hi @jsdodge ,
Thanks again for your comments!
I’ve addressed your remarks — please find my point-by-point responses below.

Looking forward to your feedback :)

---

> README
>
> This should include a short description that compares Piva to other packages in use by the ARPES community. It can be similar to what you already have in the JOSS paper.

Trying to avoid a direct reference to other packages in the README, I added the following paragraph:

*Its primary objective is to improve the efficiency of ARPES data inspection and analysis by providing interactive, intuitive tools designed to handle large, multidimensional datasets. For more information and a comparison with other packages commonly used in the ARPES community, please refer to the Statement of Need section of the accompanying manuscript.*

Please let me know if this addresses what you had in mind.

> Installation
>
> I get the following warning when running piva:
>
> piva/main.py:1: UserWarning: pkg_resources is deprecated as an
> API. See https://setuptools.pypa.io/en/latest/pkg_resources.html.
> The pkg_resources package is slated for removal as early as
> 2025-11-30. Refrain from using this package or pin to
> Setuptools<81.

I didn’t encounter the warning on my end, but I followed the recommendations provided and replaced the deprecated `pkg_resources` usage with `importlib`. This should resolve the issue.

> Tests
>
> Running pytest returned a message that it passed 4 tests and generated 21 warnings, which is different from the message given in the documentation.

I've updated the documentation to reflect the current test output.

> Data browser
>
> Documentation uses CTRL-O, macOS uses CMD-O or ⌘-O (I don't think this is a big deal, so feel free to ignore it if you think fixing it makes the documentation harder to read).

Added :)

> Data viewers
>
> The viewer axes are not labeled, nor are the values for the sliders in the various tabs. Would it be a problem to do this? It might help new users.

Axis labels were included at the very beginning, but they caused some issues. The challenge is that axis labels can vary not only across different scan types within the same dimensionality, but even within a single measurement — for example, axes may switch between kinetic and binding energy, angle and in-plane momentum, or photon energy and out-of-plane momentum. These dynamic changes led to layout shifts and misalignment of panels after each update, negatively affecting the overall display. As a result, I decided to remove the labels.

> In the image of the 2D DataViewer shown in the documentation, panel (d) has a different item order than in the actual application. The items are different in the 3D and 4D DataViewers, but their documentation tables point to the list in the 2D DataViewer table.

Corrected :)

> In the 2D DataViewer, selecting "Do it", then "Reset" changes the button name to "Do curvature".

I’ve corrected the behavior so that the button now toggles only between “Do curvature” and “Reset” as intended.

> The 3D DataViewer documentation images show "Orientate" while the application shows "Orient".

I simply formatted the images by removing the incorrect “-ate” ending.

> Selecting the "open in PIT" button in the 2D DataViewer does not issue a warning, but in the 3D DataViewer panel it does. It opens a box with the warning, "Note: PIT is a third-party package and may not work properly with Python versions above 3.8."

I had forgotten to add the warning to both viewers earlier — this is now fixed.

> The table in the 3D DataViewer documentation does not match the panels shown in the figure.

Corrected — thank you very much for spotting that!

> Selecting the "start JL session" from the "File" tab does not start JupyterLab in my installation. It does open a system file browser window, but it's not clear to me what I should select to open from this window. Files with .ipynb extensions are visible but I can not select them for opening.

The file browser is intended to let you select the root directory from which the JupyterLab server will be started. I added this feature because, once JupyterLab is running, it's not possible to navigate to parent directories — which can be frustrating when working with multiple datasets in a single session.
Just to clarify: are you saying that the server doesn’t start at all and the browser doesn’t open? Do you see any error messages when clicking the button?
At the moment I only have access to macOS, but I’ve tested the functionality thoroughly, and it should work consistently across macOS, Linux, and Windows.

> The "File" tab for the 2D DataViewer has an annotation "create experimental logbook" but the table lists "open experimental logbook" in the "Functionality" column (the description says it creates it, though).

Corrected :)

> Fitters
>
> MDC Fitter returns error when saving results:
>
> Traceback (most recent call last):
> File "/Users/jsd/Library/CloudStorage/OneDrive- SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/ fitters.py", line 1057, in save_fit_results
> for result in self.fit_results:
> TypeError: 'NoneType' object is not iterable
>
> During handling of the above exception, another exception occurred:
>
> Traceback (most recent call last):
> File "/Users/jsd/Library/CloudStorage/OneDrive- SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/ fitters.py", line 1063, in save_fit_results
> for entry in self.fit_results:
> TypeError: 'NoneType' object is not iterable

There was no safety check to verify whether any results had been added to the `self.fit_results` container. This has now been fixed — it should work as expected :)

> It's not obvious where the fit results are stored

I’ve added a message window that informs the user where the fit results were saved.

> Plot tool
>
> It is unclear to me how to use the Plot Tool.

It's true it can be a bit confusing at first, so I’ve added an additional section to the documentation that explains in more detail how to manage data curves using the Plot Tool :)
