### 2025-08-02 #231 @jsdodge
## Package Review 2

Hi @pudeIko , thanks for your response. I pulled in your changes to continue the review. Please see below for my comments on your responses. I've organized things as a task list this time.

##### README

- [X] This should include a short description that compares Piva to other packages in use by the ARPES community. It can be similar to what you already have in the JOSS paper.

> Trying to avoid a direct reference to other packages in the README, I added the following paragraph:
>
> Its primary objective is to improve the efficiency of ARPES data inspection and analysis by providing interactive, intuitive tools designed to handle large, multidimensional datasets. For more information and a comparison with other packages commonly used in the ARPES community, please refer to the Statement of Need section of the accompanying manuscript.
>
> Please let me know if this addresses what you had in mind.
>

I think that's fine.

##### Installation

- [X] I get the following warning when running `piva`:

```
piva/main.py:1: UserWarning: pkg_resources is deprecated as an
API. See https://setuptools.pypa.io/en/latest/pkg_resources.html.
The pkg_resources package is slated for removal as early as
2025-11-30. Refrain from using this package or pin to
Setuptools<81.
```

> I didn’t encounter the warning on my end, but I followed the recommendations provided and replaced the deprecated `pkg_resources` usage with `importlib`. This should resolve the issue.

This resolved the message arising from `main.py`.

- [ ] The warning persists, however, from a different module:

```
.../piva-env/lib/python3.10/site-packages/data_slicer/pit.py:11:
UserWarning: pkg_resources is deprecated as an API. See
https://setuptools.pypa.io/en/latest/pkg_resources.html. The
pkg_resources package is slated for removal as early as
2025-11-30. Refrain from using this package or pin to
Setuptools<81.
```

##### Tests

- [X] Running `pytest` returned a message that it passed 4 tests and generated 21 warnings, which is different from the message given in the [documentation](https://piva.readthedocs.io/en/latest/installation.html#testing).

> I've updated the documentation to reflect the current test output.

There's still a discrepancy for me, as shown below. The docs say 4 passed, 7 warnings. I'm not going to hold up the review over this—it's mostly for your information. Let me know if you'd like me to send you the warning information that I obtain.

`================== 4 passed, 13 warnings in 93.80s (0:01:33) ===================
`

##### Data browser

- [X] [Documentation](https://piva.readthedocs.io/en/latest/db.html) uses `CTRL-O`, macOS uses `CMD-O` or `⌘-O` (I don't think this is a big deal, so feel free to ignore it if you think fixing it makes the documentation harder to read).

##### Data viewers

- [X] The viewer axes are not labeled, nor are the values for the sliders in the various tabs. Would it be a problem to do this? It might help new users.

> Axis labels were included at the very beginning, but they caused some issues. The challenge is that axis labels can vary not only across different scan types within the same dimensionality, but even within a single measurement — for example, axes may switch between kinetic and binding energy, angle and in-plane momentum, or photon energy and out-of-plane momentum. These dynamic changes led to layout shifts and misalignment of panels after each update, negatively affecting the overall display. As a result, I decided to remove the labels.

No problem! Just wanted to check.

- [X] In the image of the 2D DataViewer shown in the [documentation](https://piva.readthedocs.io/en/latest/viewers.html#sec-utilities-panel), panel (d) has a different item order than in the actual application. The items are different in the 3D and 4D DataViewers, but their documentation tables point to the list in the 2D DataViewer table.
- [X] In the 2D DataViewer, selecting "Do it", then "Reset" changes the button name to "Do curvature".
- [X] The 3D DataViewer documentation images show "Orientate" while the application shows "Orient".
    - [x] I noticed that "Orientate" still appears in the "Axes" tab of the 3D DataViewer.
- [X] Selecting the "open in PIT" button in the 2D DataViewer does not issue a warning, but in the 3D DataViewer panel it does. It opens a box with the warning, "Note: PIT is a third-party package and may not work properly with Python versions above 3.8."
- [X] The table in the 3D DataViewer [documentation](https://piva.readthedocs.io/en/latest/viewers.html#sec-3d-viewer) does not match the panels shown in the figure.
- [x] Selecting the "start JL session" from the "File" tab does not start JupyterLab in my installation. It does open a system file browser window, but it's not clear to me what I should select to open from this window. Files with `.ipynb` extensions are visible but I can not select them for opening.

> The file browser is intended to let you select the root directory from which the JupyterLab server will be started. I added this feature because, once JupyterLab is running, it's not possible to navigate to parent directories — which can be frustrating when working with multiple datasets in a single session.
> Just to clarify: are you saying that the server doesn’t start at all and the browser doesn’t open? Do you see any error messages when clicking the button?
> At the moment I only have access to macOS, but I’ve tested the functionality thoroughly, and it should work consistently across macOS, Linux, and Windows.
>

I see now. I didn't understand the purpose of opening the file browser, so I didn't get to the point of selecting a directory. If you can add an explanatory message in the file browser window, such as "Select the root director for JupyterLab," I think that would be ideal. If not, then it would be helpful to clarify this in the table in the [File Tab](https://piva.readthedocs.io/en/latest/viewers.html#file-tab) section of the documentation.

- [X] The "File" tab for the 2D DataViewer has an annotation "create experimental logbook" but the table lists "open experimental logbook" in the "Functionality" column (the description says it creates it, though).
- [x] I neglected to mention this earlier, but it's unclear to me how to link windows in the "Volume" tab, based on the description in the table in the [Volume Tab](https://piva.readthedocs.io/en/latest/viewers.html#volume-tab) section of the documentation. Nothing happens when I select the "--select file--" drop-down menu or the "Link" button.

##### Fitters
- [X] MDC Fitter returns error when saving results:

```
  	File "/Users/jsd/Library/CloudStorage/OneDrive-	SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/	fitters.py", line 1057, in save_fit_results
   for result in self.fit_results:
	TypeError: 'NoneType' object is not iterable

  	File "/Users/jsd/Library/CloudStorage/OneDrive-	SimonFraserUniversity(1sfu)/projects/pyopensci/piva/piva/	fitters.py", line 1063, in save_fit_results
    for entry in self.fit_results:
	TypeError: 'NoneType' object is not iterable

```
- [X] It's not obvious where the fit results are stored

##### Plot tool
- [X] It is unclear to me how to use the Plot Tool.
- [ ] I played around a bit with the tool and was able to plot EDCs but not MDCs from a 2D DataViewer.
