### 2025-08-03 #231 @pudeIko
Hi @jsdodge , thanks for your prompt response! I was able to locate the issues quickly and have everything ready for your review.
Looking forward to your feedback!

---

> The warning persists, however, from a different module:
> ```
> .../piva-env/lib/python3.10/site-packages/data_slicer/pit.py:11:
> UserWarning: pkg_resources is deprecated as an API. See
> https://setuptools.pypa.io/en/latest/pkg_resources.html. The
> pkg_resources package is slated for removal as early as
> 2025-11-30. Refrain from using this package or pin to
> Setuptools<81.
> ```

I'm glad to hear that the warning isn't related to one of PIVA's modules. Unfortunately, it now appears to originate from the PIT module, which is part of the `data_slicer` package. As such, there's nothing I can do on my end to suppress it directly. However, I've added a note to the [documentation](https://piva.readthedocs.io/en/latest/viewers.html#d-dataviewer) to inform users about warnings and any potential issues that may arise.

> I noticed that "Orientate" still appears in the "Axes" tab of the 3D DataViewer.

Corrected! I ran a search, and that should be the last one 😀

> Selecting the "start JL session" from the "File" tab does not start JupyterLab in my installation. It does open a system file browser window, but it's not clear to me what I should select to open from this window. Files with .ipynb extensions are visible but I can not select them for opening.
> ```...```
> I see now. I didn't understand the purpose of opening the file browser, so I didn't get to the point of selecting a directory. If you can add an explanatory message in the file browser window, such as "Select the root director for JupyterLab," I think that would be ideal. If not, then it would be helpful to clarify this in the table in the File Tab section of the documentation.

Sure! I initially tried to set a title for the file browser window, but it doesn't show up consistently across all operating systems (notably on macOS). To ensure the message is visible regardless of platform, I added an info dialog box before the directory selection.

> I neglected to mention this earlier, but it's unclear to me how to link windows in the "Volume" tab, based on the description in the table in the Volume Tab section of the documentation. Nothing happens when I select the "--select file--" drop-down menu or the "Link" button.

To use this feature, two or more windows of the same dimensionality must be open. When clicking the dropdown menu, a list of all compatible viewers (that is having the same dimensionality, excluding the native viewer) will appear. From there, the user can select one or more viewers to link together.
Let me know if that clarifies things and whether the functionality works as expected!

> I played around a bit with the tool and was able to plot EDCs but not MDCs from a 2D DataViewer.

It was a minor bug related to axis import—but a critical one. Thanks a lot for pointing it out!
Your comment also made me rethink the design: to simplify usage of plottool, I’ve now restricted data import to viewers only, excluding fitters. This makes sense, since a 2D viewer already provides access to both sets of curves from the EDC and MDC fitters.
