### 2025-08-03 #231 @jsdodge
Hi @pudeIko , I went through your changes and I am still having trouble with plotting an MDC in the PlotTool. I'm including the steps I used to test this below. Let me know if you need more information to reproduce the issue.

1. From the Data Browser, open the `als-hers-cut.pxt` dataset from `piva/tests/data`.
2. Open the Plotting Tool
3. In the Plotting Tool, select the `als-hers-cut.pxt` dataset in the "source:"  menu of the "Add/Remove data" tab.
4. In the same tab, select `mdc` in the "plot:" menu.
5. Select `add`.
6. No data appear in the plot panel (see screenshot below).

For issues related to the "start JL session" and linking windows, I understand how to use them now and I verified both functionalities. I'd suggest that you update the descriptions in the [File Tab](https://piva.readthedocs.io/en/latest/viewers.html#file-tab) and [Volume Tab](https://piva.readthedocs.io/en/latest/viewers.html#volume-tab) sections of the documentation to clarify the points that confused me.
