### 2025-08-03 #231 @jsdodge
Yes, here it is:

```
Traceback (most recent call last):
  File "/Users/jsd/Library/CloudStorage/OneDrive-SimonFraserUniversity(1sfu)/projects/review/2025/pyopensci-piva/piva/piva/plot_tool.py", line 683, in add_dataset
    self.data_items[data_item_lbl]["data_item"] = CustomDataItem(
  File "/Users/jsd/Library/CloudStorage/OneDrive-SimonFraserUniversity(1sfu)/projects/review/2025/pyopensci-piva/piva/piva/plot_tool.py", line 1550, in __init__
    super(CustomDataItem, self).__init__(*args, **kwargs)
  File "/Users/jsd/miniforge3/envs/piva-env/lib/python3.10/site-packages/pyqtgraph/graphicsItems/PlotDataItem.py", line 375, in __init__
    self.setData(*args, **kargs)
  File "/Users/jsd/miniforge3/envs/piva-env/lib/python3.10/site-packages/pyqtgraph/graphicsItems/PlotDataItem.py", line 833, in setData
    self.updateItems( styleUpdate = self.property('styleWasChanged') )
  File "/Users/jsd/miniforge3/envs/piva-env/lib/python3.10/site-packages/pyqtgraph/graphicsItems/PlotDataItem.py", line 905, in updateItems
    self.curve.setData(x=x, y=y, **curveArgs)
  File "/Users/jsd/miniforge3/envs/piva-env/lib/python3.10/site-packages/pyqtgraph/graphicsItems/PlotCurveItem.py", line 479, in setData
    self.updateData(*args, **kargs)
  File "/Users/jsd/miniforge3/envs/piva-env/lib/python3.10/site-packages/pyqtgraph/graphicsItems/PlotCurveItem.py", line 535, in updateData
    raise Exception("X and Y arrays must be the same shape--got %s and %s." % (self.xData.shape, self.yData.shape))
Exception: X and Y arrays must be the same shape--got (368,) and (661,).
```
