### 2025-08-03 #231 @pudeIko
Ooh, that looks like an issue the latest fix should address. I cloned the repo and freshly set up the environment, and everything seems to be working properly.

Could I ask you to check whether the `get_data_from_2dviewer()` method in the `plot_tool.py` module (lines 793–802) looks like the following?

``` python
        if plot == "edc":
            if dv.new_energy_axis is None:
                return dv.data_handler.axes[1], dv.edc
            else:
                return dv.new_energy_axis, dv.edc
        elif plot == "mdc":
            if dv.k_axis is None:
                return dv.data_handler.axes[0], dv.mdc
            else:
                return dv.k_axis, dv.mdc
```

The most relevant are the indices in ```dv.data_handler.axes```.

I've also added the changes to the documentation, as requested :)
