### 2025-08-03 #231 @jsdodge
Hi @pudeIko , I'm terribly sorry, I forgot to pull in your changes before making my previous comments. Thank you for your patience. I've signed off on my review. The only item left unchecked is the citation information, which is currently listed as TBD. I'll let you sort that out with @crhea93 . Congratulations on nice work!
