### 2025-08-04 #231 @crhea93
---

🎉 PIVA has been approved by pyOpenSci! Thank you @pudeIko  for submitting PIVA and many thanks to @eigenbrot and @jsdodge  for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of PIVA. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉

---
