### 2025-08-04 #231 @crhea93
@pudeIko Congrats! I've added a comment above with the final checks/information for you. Please let me know when you complete it!

Also, if you would be interested, I invite you to submit a blog post highlighting PIVA. Please note that this is not obligatory.

If you would like to join our slack community, please let me know so that I may add you!

Since you noted that you wish to submit to JOSS in the initial issue, please let me know when you have JOSS issue created so that I can follow the progress and update this issue accordingly.

Congratulations again!!
