### 2025-08-04 #231 @pudeIko
Hi @crhea93 , @jsdodge  and @eigenbrot  !

That's fantastic news! Thank you all once again for your time, effort, and valuable feedback in helping improve the package!

@crhea93 , I'll go through the checklist in the comming days and let you know once I'm ready for the next steps :)
