### 2025-08-06 #231 @pudeIko
Hi @crhea93 ,

I’ve completed the post-review checks! 😊
The Zenodo DOI for the archived version is: 10.5281/zenodo.16755506

I released the reviewed version under the v3.0.0 tag — perhaps that could be marked as the “Version accepted”?

I’ve also submitted the paper to JOSS and will let you know as soon as the issue is opened. 😃
