### 2025-08-06 #231 @crhea93
Wonderful! I've updated the version accepted tag, added the JOSS tag, and added the archive DOI.
