### 2025-10-02 #231 @pudeIko
Hi @crhea93 ,
The package passed the first submission checks in JOSS and is now under review :)

Please find a link to the thread [here](https://github.com/openjournals/joss-reviews/issues/9129).
