### 2025-10-06 #231 @pudeIko
Hi @crhea93 ,
More great news — the manuscript has been accepted and published by JOSS! 🎉
As this marks the completion of the submission process, I’d like to once again thank you, @eigenbrot , and @jsdodge, @coatless, and @SimonMolinsky for your time and effort in helping improve the package. Best regards to you all!
